Library Reference

Graphs

AlgebraicInference.construct_elimination_sequence — Function

construct_elimination_sequence(edges::AbstractVector{T₂},
                               query::AbstractSet) where {T₁, T₂ <: AbstractSet{T₁}}

Construct an elimination sequence using the "One Step Look Ahead - Smallest Clique" heuristic.

Let $(V, E)$ be a hypergraph and $x \subseteq V$ a query. Then construct_elimination_sequence(edges, query) constructs an ordering $(X_1, \dots, X_m)$ of the vertices in $V - x$.

References:

Lehmann, N. 2001. Argumentation System and Belief Functions. Ph.D. thesis, Department of Informatics, University of Fribourg.

AlgebraicInference.construct_join_tree — Function

construct_join_tree(edges::AbstractVector{T₂},
                    elimination_sequence) where {T₁, T₂ <: AbstractSet{T₁}}

Systems

AlgebraicInference.AbstractSystem — Type

AbstractSystem

Abstract type for Gaussian systems.

Subtypes should specialize the following methods:

length(Σ::AbstractSystem)

fiber(Σ::AbstractSystem)

mean(Σ::AbstractSystem)

cov(Σ::AbstractSystem)

*(M::AbstractMatrix, Σ::AbstractSystem)

\(M::AbstractMatrix, Σ::AbstractSystem)

⊗(Σ₁::AbstractSystem, Σ₂::AbstractSystem)

References:

J. C. Willems, "Open Stochastic Systems," in IEEE Transactions on Automatic Control, vol. 58, no. 2, pp. 406-421, Feb. 2013, doi: 10.1109/TAC.2012.2210836.

AlgebraicInference.ClassicalSystem — Type

ClassicalSystem{T₁ <: AbstractMatrix, T₂ <: AbstractVector} <: AbstractSystem

A classical Gaussian system. See AbstractSystem.

AlgebraicInference.System — Type

System{T₁ <: AbstractMatrix, T₂, T₃} <: AbstractSystem

A Gaussian system. See AbstractSystem.

AlgebraicInference.ClassicalSystem — Method

ClassicalSystem(Γ::AbstractMatrix, μ::AbstractVector)

Construct a classical Gaussian system with mean $\mu$ and covariance $\Gamma$.

AlgebraicInference.ClassicalSystem — Method

ClassicalSystem(Γ::AbstractMatrix)

Construct a classical Gaussian system with mean $\mathbf{0}$ and covariance $\Gamma$.

AlgebraicInference.ClassicalSystem — Method

ClassicalSystem(μ::AbstractVector)

Construct a classical Gaussian system with mean $\mu$ and covariance $\mathbf{0}$.

AlgebraicInference.System — Method

System(R::AbstractMatrix, ϵ::ClassicalSystem)

Let $R$ be an $m \times n$ matrix and $\epsilon$ an $m$-variate random vector with mean $\mu$ and covariance $\Gamma$.

If $\mu \in \mathtt{image}(R : \Gamma)$, then there exists a random variable $\hat{w}$ taking values in $(\mathbb{R}^n, \sigma R)$ that almost-surely solves the convex program

\[ \begin{align*} \underset{w}{\text{minimize }} & E(\epsilon, w) \\ \text{subject to } & Rw \in \mathtt{image}(\Gamma) + \epsilon, \end{align*}\]

where

\[ \sigma R = \{ R^{-1}B \mid B \in \mathcal{B}(\mathbb{R}^m) \} \]

and $E(-, w)$ is the negative log-density of the multivariate normal distribution $\mathcal{N}(Rw, \Gamma)$.

If $\mu \in \mathtt{image}(R : \Gamma)$, then System(R, ϵ) constructs the Gaussian system $\Sigma = (\mathbb{R}^n, \sigma R, P)$, where $P$ is the distribution of $\hat{w}$.

In particular, if $R$ has full row-rank, then $Rw = \epsilon$ is a kernel representation of $\Sigma$.

AlgebraicInference.System — Method

System(R::AbstractMatrix)

Let $R$ be an $m \times n$ matrix. Then System(R) constructs the deterministic Gaussian system $\Sigma = (\mathbb{R}^n, \sigma R, P),$ where

\[ \sigma R = \{ R^{-1}B \mid B \in \mathcal{B}(\mathbb{R}^m)\}\]

and

\[ P(R^{-1}B) = \begin{cases} 1 & 0 \in B \\ 0 & \text{else} \end{cases}.\]

Base.length — Method

length(Σ::AbstractSystem)

Let $\Sigma = (\mathbb{R}^n, \mathcal{E}, P)$. Then length(Σ) gets the dimension $n$.

StatsAPI.dof — Method

dof(Σ::AbstractSystem)

Get the number of degrees of freedom of $Σ$.

AlgebraicInference.fiber — Method

fiber(Σ::AbstractSystem)

Compute a basis for the fiber of $Σ$.

Statistics.mean — Method

mean(Σ::AbstractSystem)

Let $Rw = \epsilon$ be any kernel representation of $\Sigma$. Then mean(Σ) computes a vector $\mu$ such that $R\mu$ is the mean of $\epsilon$.

In particular, if $\Sigma$ is a classical Gaussian system, then $\mu$ is the mean of $\Sigma$.

Statistics.cov — Method

cov(Σ::AbstractSystem)

Let $Rw = \epsilon$ be any kernel representation of $\Sigma$. Then cov(Σ) computes a matrix $\Gamma$ such that $R \Gamma R^\mathsf{T}$ is the covariance of $\epsilon$.

In particular, if $\Sigma$ is a classical Gaussian system, then $\Gamma$ is the covariance of $\Sigma$.

Base.:* — Method

*(M::AbstractMatrix, Σ::AbstractSystem)

Let $M$ be an $n \times m$ matrix, and let $\Sigma = (\mathbb{R}^m, \mathcal{E}, P)$. Then M * Σ computes the Gaussian system $\Sigma' = (\mathbb{R}^n, \mathcal{E}', P')$, where

\[ \mathcal{E}' = \{ B \in \mathcal{B}(\mathbb{R}^n) \mid M^{-1}B \in \mathcal{E} \}\]

and

\[ P'(B) = P(M^{-1}B).\]

Base.:\ — Method

\(M::AbstractMatrix, Σ::AbstractSystem)

Catlab.Theories.:⊗ — Method

⊗(Σ₁::AbstractSystem, Σ₂::AbstractSystem)

Compute the product $\Sigma_1 \times \Sigma_2$.

Catlab.WiringDiagrams.WiringDiagramAlgebras.oapply — Method

oapply(composite::UndirectedWiringDiagram,
       box_map::AbstractDict{T₁, T₂}) where {T₁, T₂ <: AbstractSystem}

Catlab.WiringDiagrams.WiringDiagramAlgebras.oapply — Method

oapply(composite::UndirectedWiringDiagram,
       boxes::AbstractVector{T}) where T <: AbstractSystem

Valuations

AlgebraicInference.Variable — Type

Variable

Abtract type for variables in a valuation algebra. See Valuation.

AlgebraicInference.Valuation — Type

Valuation{T <: Variable}

Abstract type for valuations in a valuation algebra. For any type T <: Variable, the types Valuation{T} and T should form a stable valuation algebra.

Subtypes should specialize the following methods:

domain(ϕ::Valuation)

combine(ϕ₁::Valuation{T}, ϕ₂::Valuation{T}) where T

project(ϕ::Valuation{T}, x::AbstractSet{T}) where T

neutral_valuation(x::AbstractSet{T}) where T <: Variable

References:

Pouly, M.; Kohlas, J. Generic Inference. A Unified Theory for Automated Reasoning; Wiley: Hoboken, NJ, USA, 2011.

AlgebraicInference.LabeledBoxVariable — Type

LabeledBoxVariable{T} <: Variable

AlgebraicInference.IdentityValuation — Type

IdentityValuation{T} <: Valuation{T}

The identity element $e$.

AlgebraicInference.LabeledBox — Type

LabeledBox{T₁, T₂} <: Valuation{LabeledBoxVariable{T₁}}

AlgebraicInference.domain — Function

domain(ϕ::Valuation)

Get the domain of $\phi$.

AlgebraicInference.combine — Function

combine(ϕ₁::Valuation{T}, ϕ₂::Valuation{T}) where T

Perform the combination $\phi_1 \otimes \phi_2$.

AlgebraicInference.project — Function

project(ϕ::Valuation{T}, x::AbstractSet{T}) where T

Perform the projection $\phi^{\downarrow x}$.

AlgebraicInference.neutral_valuation — Function

neutral_valuation(x::AbstractSet{T}) where T <: Variable

Construct the neutral element $\phi^{\downarrow x}$.

AlgebraicInference.eliminate — Function

eliminate(ϕ::Valuation{T}, X::T) where T <: Variable

Perform the variable elimination $\phi^{-X}$.

AlgebraicInference.construct_inference_problem — Method

construct_inference_problem(::Type,
                            composite::UndirectedWiringDiagram,
                            box_map::AbstractDict) where T

AlgebraicInference.construct_inference_problem — Method

construct_inference_problem(::Type,
                            composite::UndirectedWiringDiagram,
                            boxes::AbstractVector)

Let $f$ be an operation in Cospan of the form

\[ B \xleftarrow{\mathtt{box}} P \xrightarrow{\mathtt{junc}} J \xleftarrow{\mathtt{junc'}} Q,\]

where $\mathtt{junc'}: Q \to J$ is injective, and $(b_1, \dots, b_n)$ a sequence of fillers for the boxes in $f$. Then construct_inference_problem(T, composite, boxes) constructs a knowledge base $\{\phi_1, \dots, \phi_n\}$ and query $x \subseteq J$ such that

\[ (\phi_1 \otimes \dots \otimes \phi_n)^{\downarrow x} \cong F(f)(b_1, \dots, b_n),\]

where $F$ is the Cospan-algebra computed by oapply.

AlgebraicInference.construct_factors — Function

construct_factors(knowledge_base::AbstractVector{<:Valuation{T₁}},
                  assignment_map::AbstractVector{Int},
                  domains::AbstractVector{T₂},
                  tree::Node{Int};
                  identity=false) where {T₁ <: Variable, T₂ <: AbstractSet{T₁}}

AlgebraicInference.fusion_algorithm — Function

fusion_algorithm(knowledge_base::AbstractVector{<:Valuation{T}},
                 elimination_sequence::AbstractVector{T}) where T <: Variable

An implementation of the fusion algorithm.

AlgebraicInference.collect_algorithm — Function

collect_algorithm(factors::AbstractVector{<:Valuation{T₁}},
                  domains::AbstractVector{T₂},
                  tree::Node{Int},
                  query::AbstractSet{T₁}) where {T₁ <: Variable, T₂ <: AbstractSet{T₁}}

An implementation of the collect algorithm.

AlgebraicInference.shenoy_shafer_architecture! — Function

shenoy_shafer_architecture!(mailboxes::AbstractDict{Tuple{Int, Int}, Valuation{T₁}},
                            factors::AbstractVector{<:Valuation{T₁}},
                            domains::AbstractVector{T₂},
                            tree::Node{Int},
                            query::AbstractSet{T₁}) where {T₁ <: Variable, T₂ <: AbstractSet{T₁}}

An implementation of the Shenoy-Shafer architecture.