Library Reference

AlgebraicInference.GaussianDistribution — Type

GaussianDistribution(Σ, μ)

A multivariate Gaussian distribution with covariance matrix $\Sigma$ and mean $\mu$.

AlgebraicInference.GaussianDistribution — Method

GaussianDistribution(Σ::AbstractMatrix)

Construct a centered Gaussian distribution with covariance matrix $\Sigma$.

AlgebraicInference.GaussianDistribution — Method

GaussianDistribution(μ::AbstractVector)

Construct a Dirac distribution with mean $\mu$.

AlgebraicInference.GaussianRelation — Type

GaussianRelation(L, R, ψ)

A Gaussian relation, i.e. a decorated cospan with legs L, R and decoration ψ.

AlgebraicInference.GaussianRelation — Method

GaussianRelation(L::AbstractMatrix)

Construct a Gaussian relation $d: m \to n$ from a linear transformation $L: \mathbb{R}^m \to \mathbb{R}^n$.

AlgebraicInference.GaussianRelation — Method

GaussianRelation(ψ::GaussianDistribution)

Construct a Gaussian relation $d: 0 \to n$ from a Gaussian distribution $\psi$ on $\mathbb{R}^n$.

AlgebraicInference.GaussRelDom — Type

GaussRelDom(n)

The domain or codomain of a Gaussian relation.

StatsAPI.params — Function

params(d::GaussianRelation)

Represent a Gaussian relation $d: m \to n$ as an extended Gaussian distribution $\mathcal{N}(\mu, \Sigma) + \text{null } D$ on $\mathbb{R}^{m + n}$.

Returns a five-tuple $(Σ, μ, D, B, b)$. If $b \notin \text{col } B$, then $d$ is equal to the failure morphism $\bot: m \to n$.

Statistics.cov — Function

cov(ψ::GaussianDistribution)

Return the covariance matrix corresponding to ψ.

cov(d::GaussianRelation)

Return a representative of the covariance matrix corresponding to d.

Statistics.mean — Function

mean(ψ::GaussianDistributon)

Return the mean vector corresponding to ψ.

mean(d::GaussianRelation)

Return a representative of the mean vector corresponding to d.