Library Reference

Architectures

AlgebraicInference.Architecture — Type

Architecture{T₁, T₂} <: AbstractNode{T₁}

A join tree $(V, E, \lambda, D)$, along with a set of factors

\[ \left\{ \phi_i \right\}_{i \in V}\]

and mailboxes

\[ \left\{ \left( \mu_{ i \to \mathtt{pa}(i)}, \mu_{\mathtt{pa}(i) \to i} \right) \right\}_{i \in V}.\]

AlgebraicInference.architecture — Function

architecture(kb::AbstractVector{<:Valuation{T}}, order) where T

Construct a covering join tree for the knowledge base kb using the variable elimination order order.

AlgebraicInference.answer_query — Function

answer_query(jt::Architecture{T₁, T₂}, query::Set{T₂}) where {T₁, T₂}

Answer a query.

Let $(V, E, \lambda, D)$ be a join tree with factors $\{\phi_i\}_{i \in V}$ and $x$ a query covered by $(V, E, \lambda, D)$. Then answer_query(jt, query) solves the inference problem

\[\left( \bigotimes_{i \in V} \psi_i \right)^{\downarrow x}.\]

AlgebraicInference.answer_query! — Function

answer_query!(jt::Architecture{T₁, T₂}, query::Set{T₂}) where {T₁, T₂}

Answer a query, caching intermediate computations.

Let $(V, E, \lambda, D)$ be a join tree with factors $\{\phi_i\}_{i \in V}$ and $x$ be a query covered by $(V, E, \lambda, D)$. Then answer_query!(jt, query) solves the inference problem

\[\left( \bigotimes_{i \in V} \psi_i \right)^{\downarrow x}.\]

Graphs

AlgebraicInference.primal_graph — Function

primal_graph(kb::AbstractVector{<:Valuation{T}}) where T

Construct the primal graph of the knowledge base kb.

AlgebraicInference.minfill! — Function

minfill!(g::MetaGraph, query)

Compute a vertex elimination order using the min-fill heuristic.

AlgebraicInference.minwidth! — Function

minwidth!(g::MetaGraph, query)

Compute a vertex elimination order using the min-width heuristic

Systems

AlgebraicInference.GaussianSystem — Type

GaussianSystem{
    T₁ <: AbstractMatrix,
    T₂ <: AbstractMatrix, 
    T₃ <: AbstractVector,
    T₄ <: AbstractVector,
    T₅}

A Gaussian system.

References:

J. C. Willems, "Open Stochastic Systems," in IEEE Transactions on Automatic Control, vol. 58, no. 2, pp. 406-421, Feb. 2013, doi: 10.1109/TAC.2012.2210836.

AlgebraicInference.canon — Method

canon(J::AbstractMatrix, h::AbstractVector)

Construct a multivariate nornal distribution with information matrix J and potential vector h.

AlgebraicInference.canon — Method

canon(J::AbstractMatrix)

Construct a centered multivariate normal distribution with information matrix J.

AlgebraicInference.normal — Method

normal(Σ::AbstractMatrix, μ::AbstractVector)

Construct a multivariate normal distribution with covariance matrix Σ and mean vector μ.

AlgebraicInference.normal — Method

normal(Σ::AbstractMatrix)

Construct a centered multivariate normal distribution with covariance matrix Σ.

AlgebraicInference.normal — Method

normal(μ::AbstractVector)

Construct a Dirac distribution with mean vector μ.

AlgebraicInference.kernel — Method

kernel(Σ::AbstractMatrix, μ::AbstractVector, L::AbstractMatrix)

Construct a conditional distribution of the form

\[ y \mid x \sim \mathcal{N}(Lx + \mu, \Sigma).\]

AlgebraicInference.kernel — Method

kernel(Σ::AbstractMatrix, L::AbstractMatrix)

Construct a conditional distribution of the form

\[ y \mid x \sim \mathcal{N}(Lx, \Sigma).\]

AlgebraicInference.kernel — Method

kernel(L::AbstractMatrix)

Construct a conditional distribution of the form

\[ y \mid x \sim \delta_{Lx}.\]

Base.length — Method

length(Σ::GaussianSystem)

Get the dimension of Σ.

Statistics.cov — Method

cov(Σ::GaussianSystem)

Get the covariance matrix of Σ.

AlgebraicInference.invcov — Method

invcov(Σ::GaussianSystem)

Get the information matrix of Σ.

Statistics.mean — Method

mean(Σ::GaussianSystem)

Get the mean vector of Σ.

Catlab.WiringDiagrams.WiringDiagramAlgebras.oapply — Method

oapply(wd::UndirectedWiringDiagram, box_map::AbstractDict{<:Any, <:GaussianSystem})

See oapply(wd::UndirectedWiringDiagram, box_map::AbstractVector{<:GaussianSystem}).

Catlab.WiringDiagrams.WiringDiagramAlgebras.oapply — Method

oapply(wd::UndirectedWiringDiagram, boxes::AbstractVector{<:GaussianSystem})

Compose Gaussian systems according to the undirected wiring diagram wd.

Valuations

AlgebraicInference.Valuation — Type

Valuation{T}

Abstract type for valuations in a valuation algebra.

Subtypes should specialize the following methods:

domain(ϕ::Valuation)

combine(ϕ₁::Valuation, ϕ₂::Valuation)

project(ϕ::Valuation, x)

Valuations are parametrized by the type of the variables in their variable system. If isa(ϕ, Valuation{T}), then domain(ϕ) should return a container with element type T.

References:

Pouly, M.; Kohlas, J. Generic Inference. A Unified Theory for Automated Reasoning; Wiley: Hoboken, NJ, USA, 2011.

AlgebraicInference.IdentityValuation — Type

IdentityValuation{T} <: Valuation{T}

The identity element $e$.

AlgebraicInference.LabeledBox — Type

LabeledBox{T₁, T₂} <: Valuation{T₁}

AlgebraicInference.domain — Function

domain(ϕ::Valuation)

Get the domain of $\phi$.

AlgebraicInference.combine — Function

combine(ϕ₁::Valuation, ϕ₂::Valuation)

Perform the combination $\phi_1 \otimes \phi_2$.

AlgebraicInference.project — Function

project(ϕ::Valuation, x)

Perform the projection $\phi^{\downarrow x}$.

AlgebraicInference.inference_problem — Method

inference_problem(wd::UndirectedWiringDiagram, box_map::AbstractDict)

See inference_problem(wd::UndirectedWiringDiagram, boxes::AbstractVector).

AlgebraicInference.inference_problem — Method

inference_problem(wd::UndirectedWiringDiagram, boxes::AbstractVector)

Let $f$ be an operation in Cospan of the form

\[ B \xleftarrow{\mathtt{box}} P \xrightarrow{\mathtt{junc}} J \xleftarrow{\mathtt{junc'}} Q\]

and $(b_1, \dots, b_n)$ a sequence of fillers for the boxes in $f$. Then inference_problem(wd, boxes) constructs a knowledge base $\{\phi_1, \dots, \phi_n\}$ and query $x \subseteq J$ such that

\[ (\phi_1 \otimes \dots \otimes \phi_n)^{\downarrow x} \cong F(f)(b_1, \dots, b_n),\]

where $F$ is the Cospan-algebra computed by oapply.

The operation $f$ must satify must satisfy the following constraints:

$\mathtt{junc'}$ is injective.

$\mathtt{image}(\mathtt{junc'}) \subseteq \mathtt{image}(\mathtt{junc})$

For all $x, y \in P$, $\mathtt{box}(x) = \mathtt{box}(y)$ and $\mathtt{junc}(x) = \mathtt{junc}(y)$ implies that $x = y$.