Library Reference

Systems

AlgebraicInference.GaussianSystem — Type

GaussianSystem{T₁, T₂, T₃, T₄, T₅}

A Gaussian system.

AlgebraicInference.GaussianSystem — Method

GaussianSystem{T₁, T₂, T₃, T₄, T₅}(P, S, p, s, σ) where {
    T₁ <: AbstractMatrix,
    T₂ <: AbstractMatrix,
    T₃ <: AbstractVector,
    T₄ <: AbstractVector,
    T₅ <: Real}

Construct a Gaussian system by specifying its energy function.

You should set σ equal to $s^\mathsf{T} S^+ s$, where $S^+$ is the Moore-Penrose psuedoinverse of $S$.

GaussianSystem(
    P::AbstractMatrix,
    S::AbstractMatrix,
    p::AbstractVector,
    s::AbstractVector,
    σ::Real)

Construct a Gaussian system by specifying its energy function.

You should set σ equal to $s^\mathsf{T} S^+ s$, where $S^+$ is the Moore-Penrose psuedoinverse of $S$.

AlgebraicInference.normal — Function

normal(μ::AbstractVector, Σ::AbstractMatrix)

Construct a multivariate normal distribution with mean vector μ and covariance matrix Σ.

normal(μ::Real, σ::Real)

Construct a normal distribution with mean μ and standard deviation σ.

AlgebraicInference.kernel — Function

kernel(L::AbstractMatrix, μ::AbstractVector, Σ::AbstractMatrix)

Construct a conditional distribution of the form $(y \mid x) \sim \mathcal{N}(Lx + \mu, \Sigma).$

kernel(l::AbstractVector, μ::Real, σ::Real)

Construct a conditional distribution of the form $(y \mid x) \sim \mathcal{N}(l^\mathsf{T}x + \mu, \sigma^2).$

Base.length — Method

length(Σ::GaussianSystem)

Get the dimension of Σ.

Statistics.cov — Method

cov(Σ::GaussianSystem; atol=1e-8)

Get the covariance matrix of Σ.

Distributions.invcov — Method

invcov(Σ::GaussianSystem)

Get the precision matrix of Σ.

Statistics.var — Method

var(Σ::GaussianSystem; atol=1e-8)

Get the variances of Σ.

Statistics.mean — Method

mean(Σ::GaussianSystem; atol=1e-8)

Get the mean vector of Σ.

Catlab.WiringDiagrams.WiringDiagramAlgebras.oapply — Method

oapply(wd::AbstractUWD, homs::AbstractVector{<:GaussianSystem}, obs::AbstractVector)

Compose Gaussian systems according to the undirected wiring diagram wd.

Problems

AlgebraicInference.InferenceProblem — Type

InferenceProblem{T₁, T₂, T₃, T₄}

An inference problem computes the conditional distribution of $X$ given $Y = e$, where $X$ and $Y$ are random variables whose joint probability distribution is specified by a graphical model.

AlgebraicInference.InferenceProblem — Method

InferenceProblem{T₁, T₂, T₃, T₄}(
    wd::AbstractUWD,
    hom_map::AbstractDict,
    ob_map::AbstractDict;
    hom_attr::Symbol=:name,
    ob_attr::Symbol=:variable) where {T₁, T₂, T₃, T₄}

Construct an inference problem that performs undirected compositon.

AlgebraicInference.InferenceProblem — Method

InferenceProblem{T₁, T₂, T₃, T₄}(
    wd::AbstractUWD,
    homs::AbstractVector,
    obs::AbstractVector) where {T₁, T₂, T₃, T₄}

Construct an inference problem that performs undirected compositon.

AlgebraicInference.InferenceProblem — Method

InferenceProblem{T₁, T₂, T₃, T₄}(
    bn::BayesNet,
    query::AbstractVector,
    evidence::AbstractDict) where {T₁, T₂, T₃, T₄}

CommonSolve.solve — Method

solve(ip::InferenceProblem, alg::EliminationAlgorithm)

Solve an inference problem.

CommonSolve.init — Method

init(ip::InferenceProblem, alg::EliminationAlgorithm)

Construct a solver for an inference problem.

Solvers

AlgebraicInference.InferenceSolver — Type

InferenceSolver{T₁, T₂, T₃}

A solver for an inference problem.

An InferenceSolver can be reused to answer multiple queries:

is = init(ip)
sol₁ = solve(is)
is.query = query
sol₂ = solve(is)

CommonSolve.solve — Method

solve(is::InferenceSolver)

Solve an inference problem.

CommonSolve.solve! — Method

solve!(is::InferenceSolver)

Solve an inference problem, caching intermediate computations.

Algorithms

AlgebraicInference.EliminationAlgorithm — Type

EliminationAlgorithm

An algorithm for ordering the vertices of an undirected graph. The ordering is used to construct a join tree.

AlgebraicInference.MinDegree — Type

MinDegree <: EliminationAlgorithm

The min-degree heuristic.

AlgebraicInference.MinFill — Type

MinFill <: EliminationAlgorithm

The min-fill heuristic.