Library Reference

Gaussian Distributions

AlgebraicInference.GaussianDistribution — Type

GaussianDistribution(cgf)

A multivariate Gaussian distribution.

AlgebraicInference.GaussianDistribution — Method

GaussianDistribution(Σ::AbstractMatrix, μ::AbstractVector)

Construct a Gaussian distribution with covariance Σ and mean μ.

AlgebraicInference.GaussianDistribution — Method

GaussianDistribution(Σ::AbstractMatrix)

Construct a centered Gaussian distribution with covariance Σ.

AlgebraicInference.GaussianDistribution — Method

GaussianDistribution(μ::AbstractVector)

Construct a Dirac distribution with mean μ.

AlgebraicInference.OpenGaussianDistribution — Type

OpenGaussianDistribution(rf)

An open Gaussian distribution is a decorated cospan

\[\left( \begin{aligned} \begin{CD} m @>L>> k @<R<< n \end{CD} \end{aligned}, \enspace \begin{aligned} \mathcal{N}(\mu, \Sigma) \end{aligned} \right),\]

where $L$, $R$ are matrices and $\mathcal{N}(\mu, \Sigma)$ is a Gaussian distribution on $\mathbb{R}^k$. This data determines a morphism

\[d: m \to n\]

in the hypergraph category $\text{Cond}(\text{GaussEx})$. In particular, if $m = 0$, then $d$ represents the posterior distribution

\[X \mid Y = 0\]

of a Bayesian model

\[\begin{align*} X &\sim \mathcal{U} \\ \epsilon &\sim \mathcal{N}(\mu, \Sigma) \\ Y &= RX - \epsilon, \end{align*}\]

where $\mathcal{U}$ is an improper uniform prior over $\mathbb{R}^n$ and $X \perp \epsilon$.

References:

Stein & Staton (2021), "Compositional Semantics for Probabilistic Programs with Exact Conditioning" (arXiv:2101.11351)

Stein (2022), "Decorated Linear Relations: Extending Gaussian Probability with Uninformative Priors" (arXiv:2204.14024)

Stein (2022), "A Hypergraph Category for Exact Gaussian Inference", (https://msp.cis.strath.ac.uk/act2022/papers/ACT2022paper3601.pdf)

AlgebraicInference.OpenGaussianDistribution — Method

OpenGaussianDistribution(L::AbstractMatrix, R::AbstractMatrix, ψ::GaussianDistribution)

Construct an open Gaussian distribution with legs (L, R) and decoration ψ.

AlgebraicInference.OpenGaussianDistribution — Method

OpenGaussianDistribution(L::AbstractMatrix, ψ::GaussianDistribution)

Construct the Gaussian map

\[x \mapsto Lx + \psi\]

AlgebraicInference.OpenGaussianDistribution — Method

OpenGaussianDistribution(L::AbstractMatrix)

Construct the linear map

\[x \mapsto Lx\]

AlgebraicInference.OpenGaussianDistribution — Method

OpenGaussianDistribution(ψ::GaussianDistribution)

Construct the Gaussian map

\[* \mapsto \psi\]

AlgebraicInference.GaussDom — Type

GaussDom(n)

The Euclidean space $\mathbb{R}^n$.

StatsAPI.params — Function

params(d::OpenGaussianDistribution)

Compute the cumulant-generating function $K$ of d.

Returns a quadruple $(Q, a, B, b)$. If $b \neq 0$, then $K = -\infty$. Otherwise,

\[K(x, y) = \begin{cases} \langle (x, y), \frac{1}{2}Q(x, y) + a \rangle & B(x, y) = 0 \\ \infty & \text{else} \end{cases}.\]

Statistics.cov — Function

cov(ψ::GaussianDistribution)

Get the covariance of ψ.

cov(d::OpenGaussianDistribution)

Get the covariance of d.

Statistics.mean — Function

mean(ψ::GaussianDistribution)

Get the mean of ψ.

mean(d::OpenGaussianDistribution)

Get the mean of d.

Quadratic Functions

AlgebraicInference.QuadraticFunction — Type

QuadraticFunction(Q, a)

A convex quadratic function of the form

\[f(x) = \langle x, Qx + a \rangle,\]

where Q is positive semidefinite.

AlgebraicInference.QuadraticFunction — Method

QuadraticFunction(Q::AbstractMatrix)

Construct the convex quadratic function

\[f(x) = \langle x, Qx \rangle,\]

where $Q$ is positive semidefinite.

AlgebraicInference.QuadraticFunction — Method

QuadraticFunction(a::AbstractVector)

Construct the linear function

\[f(x) = \langle x, a \rangle.\]

AlgebraicInference.OpenQuadraticFunction — Type

OpenQuadraticFunction(L, R, f)

An open quadratic function is a decorated cospan

\[\left( \begin{aligned} \begin{CD} m @>L>> k @<R<< n \end{CD} \end{aligned}, \enspace \begin{aligned} f \end{aligned} \right),\]

where $L$, $R$ are matrices and $f: \mathbb{R}^k \to \mathbb{R}$ is a convex quadratic function. This data determines a convex bifunction

\[F(x, y) = f^*(Ry - Lx),\]

where $f^*$ is the convex conjuate of $f$.

AlgebraicInference.OpenQuadraticFunction — Method

OpenQuadraticFunction(L::AbstractMatrix, f::QuadraticFunction)

Construct the bifunction

\[F(x, y) = f^*(y - Lx),\]

where $f^*$ is the convex conjugate of $f$.

AlgebraicInference.OpenQuadraticFunction — Method

OpenQuadraticFunction(L::AbstractMatrix)

Construct the indicator bifunction corresponding to L, given by

\[F(x, y) = \begin{cases} 0 & Lx = y \\ \infty & \text{else} \end{cases}.\]

AlgebraicInference.OpenQuadraticFunction — Method

OpenQuadraticFunction(f::QuadraticFunction)

Construct the bifunction

\[F(*, y) = f^*(y),\]

where $f^*$ is the convex conjugate of $f$.

AlgebraicInference.QuadDom — Type

QuadDom(n)

The Euclidian space $\mathbb{R}^n$.

Base.:* — Function

*(f::QuadraticFunction, M::Union{UniformScaling, AbstractMatrix})

Compute the inverse image of f under M, given by

\[(fM)(y) = f(My).\]

*(M::Union{UniformScaling, AbstractMatrix}, f::QuadraticFunction)

Compute the image of f under M, given by

\[(Mf)(y) = \inf \{ f(x) \mid y = Mx \}.\]

Returns a quintuple $(Q, a, \alpha, B, b).$ If $b \neq 0$, then $Mf = -\infty$. Otherwise,

\[(Mf)(y) = \begin{cases} \langle y, \frac{1}{2}Qy + a \rangle + \alpha & By = 0 \\ \infty & \text{else} \end{cases}.\]

AlgebraicInference.conjugate — Function

conjugate(F::OpenQuadraticFunction)

Compute the convex conjugate of F.

Returns a quintuple $(Q, a, \alpha, B, b).$ If $b \neq 0$, then $F^* = -\infty$. Otherwise,

\[F^*(x, y) = \begin{cases} \langle (x, y), \frac{1}{2}Q(x, y) + a \rangle + \alpha & B(x, y) = 0 \\ \infty & \text{else} \end{cases}.\]

Composition

Catlab.Theories.compose — Function

compose(F₁::OpenQuadraticFunction, F₂::OpenQuadraticFunction)

Compute the product of F₁ and F₂, given by

\[(F_2F_1)(x, z) = \inf \{ F_1(x, y) + F_2(y, z) \mid y \}.\]

compose(d₁::OpenGaussianDistribution, d₂::OpenGaussianDistribution)

Catlab.Theories.oplus — Function

oplus(f₁::QuadraticFunction, f₂::QuadraticFunction)

Compute the direct sum of f₁ and f₂, given by

\[(f_1 \oplus f_2)(x, y) = f_1(x) + f_2(y).\]

oplus(X::QuadDom, Y::QuadDom)

Compute the direct sum of $X$ and $Y$.

oplus(F₁::OpenQuadraticFunction, F₂::OpenQuadraticFunction)

Compute the direct sum of F₁ and F₂, given by

\[(F_1 \oplus F_2)((x_1, x_2), (y_1, y_2)) = F_1(x_1, y_1) + F_2(x_2, y_2).\]

oplus(X::GaussDom, Y::GassDom)

Compute direct sum of X and Y.

oplus(d₁::OpenGaussianDistribution, d₂::OpenGaussianDistribution)

Catlab.Theories.meet — Function

meet(F₁::OpenQuadraticFunction, F₂::OpenQuadraticFunction)

Compute the sum of F₁ and F₂, given by

\[(F_1 + F_2)(x, y) = F_1(x, y) + F_2(x, y).\]

meet(d₁::OpenGaussianDistribution, d₂::OpenGaussianDistribution)

Base.join — Function

join(F₁::OpenQuadraticFunction, F₂::OpenQuadraticFunction)

Compute the infimal convolution of F₁ and F₂, given by

\[(F_1 \, \Box \, F_2)(x^*, y^*) = \inf \{ F_1(x^*, y^*) + F_2(x - x^*, y - y^*) \mid x, y \}.\]

join(d₁::OpenGaussianDistribution, d₂::OpenGaussianDistribution)

Catlab.WiringDiagrams.WiringDiagramAlgebras.oapply — Function

oapply(composite::UndirectedWiringDiagram, hom_map::AbstractDict{T₁, T₂}) where {T₁, T₂ <: OpenQuadraticFunction}

Compose open quadratic functions according to an undirected wiring wiring diagram.

oapply(composite::UndirectedWiringDiagram, cospans::AbstractVector{T}) where T <: OpenQuadraticFunction

Compose open quadratic functions according to an undirected wiring wiring diagram.

oapply(composite::UndirectedWiringDiagram, hom_map::AbstractDict{T₁, T₂}) where {T₁, T₂ <: OpenGaussianDistribution}

Compose open Gaussian distributions according to an undirected wiring wiring diagram.

oapply(composite::UndirectedWiringDiagram, cospans::AbstractVector{T}) where T <: OpenGaussianDistribution

Compose open Gaussian distributions according to an undirected wiring wiring diagram.

Construction

Catlab.Theories.mzero — Function

mzero(::Type{QuadDom})

Construct the Euclidean space $\mathbb{R}^0$.

mzero(::Type{GaussDom})

Construct the Euclidean space $\mathbb{R}^0$.

Catlab.Theories.id — Function

id(X::QuadDom)

Construct the indicator bifunction

\[F(x, y) = \begin{cases} 0 & x = y \\ \infty & \text{else} \end{cases}.\]

id(X::GaussDom)

Catlab.Theories.swap — Function

swap(X::QuadDom, Y::QuadDom)

Construct the indicator bifunction

\[F((x_1, x_2), (y_1, y_2)) = \begin{cases} 0 & x_1 = y_2, x_2 = y_1 \\ \infty & \text{else} \end{cases}.\]

swap(X::GaussDom, Y::GaussDom)

Construct the linear map

\[(x_1, x_2) \mapsto (x_2, x_1).\]

Base.zero — Function

zero(X::QuadDom)

Construct the indicator bifunction

\[F(*, y) = \begin{cases} 0 & y = 0 \\ \infty & \text{else} \end{cases}\]

zero(X::GaussDom)

Construct the linear map

\[* \mapsto 0.\]

Catlab.Theories.cozero — Function

cozero(X::QuadDom)

Construct the indicator bifunction

\[F(x, *) = \begin{cases} 0 & x = 0 \\ \infty & \text{else} \end{cases}.\]

cozero(X::GaussDom)

Catlab.Theories.delete — Function

delete(X::QuadDom)

Construct the indicator bifunction

\[F(x, *) = 0.\]

delete(X::GaussDom)

Construct the linear map

\[x \mapsto *.\]

Catlab.Theories.create — Function

create(X::QuadDom)

Construct the indicator bifunction

\[F(*, y) = 0.\]

create(X::GaussDom)

Construct the extended Gaussian map

\[* \mapsto \mathbb{R}^n.\]

Catlab.Theories.mcopy — Function

mcopy(X::QuadDom)

Construct the indicator bifunction

\[F(x, (y_1, y_2)) = \begin{cases} 0 & x = y_1 = y_2 \\ \infty & \text{else} \end{cases}.\]

mcopy(X::GaussDom)

Construct the linear map

\[x \mapsto (x, x).\]

Catlab.Theories.mmerge — Function

mmerge(X::QuadDom)

Construct the indicator bifunction

\[F((x_1, x_2), y) = \begin{cases} 0 & x_1 = x_2 = y \\ \infty & \text{else} \end{cases}.\]

mmerge(X::GaussDom)

Catlab.Theories.plus — Function

plus(X::QuadDom)

Construct the indicator bifunction

\[F((x_1, x_2), y) = \begin{cases} 0 & x_1 + x_2 = y \\ \infty & \text{else} \end{cases}.\]

plus(X::GaussDom)

Construct the linear map

\[(x, y) \mapsto x + y.\]

Catlab.Theories.coplus — Function

coplus(X::QuadDom)

Construct the indicator bifunction

\[F((x_1, x_2), y) = \begin{cases} 0 & x_1 + x_2 = y \\ \infty & \text{else} \end{cases}\]

coplus(X::GaussDom)

Catlab.Theories.dunit — Function

dunit(X::QuadDom)

Construct the indicator bifunction

\[F(*, (y_1, y_2)) = \begin{cases} 0 & y_1 = y_2 \\ \infty & \text{else} \end{cases}.\]

dunit(X::GaussDom)

Construct the extended Gaussian map

\[* \mapsto \{(y, y) \mid y \}.\]

Catlab.Theories.dcounit — Function

dcounit(X::QuadDom)

Construct the indicator bifunction

\[F((x_1, x_2), *) = \begin{cases} 0 & x_1 = x_2 \\ \infty & \text{else} \end{cases}.\]

dcounit(X::GaussDom)

Catlab.Theories.top — Function

top(X::QuadDom, Y::QuadDom)

Construct the indicator bifunction

\[F(x, y) = 0.\]

top(X::GaussDom, Y::GaussDom)

Construct the extended Gaussian map

\[x \mapsto \mathbb{R}^n.\]

Catlab.Theories.bottom — Function

bottom(X::QuadDom, Y::QuadDom)

Construct the indicator bifunction

\[F(x, y) = \begin{cases} 0 & x = y = 0 \\ \infty & \text{else} \end{cases}.\]

bottom(X::GaussDom, Y::GaussDom)

Other

Catlab.Schemas.dom — Function

dom(F::OpenQuadraticFunction)

The domain of a bifunction is the dimension of its first variable.

dom(d::OpenGaussianDistribution)

Catlab.Schemas.codom — Function

codom(F::OpenQuadraticFunction)

The codomain of a bifunction is the dimension of its second variable.

codom(d::OpenGaussianDistribution)

Catlab.Theories.dagger — Function

dagger(F::OpenQuadraticFunction)

Compute the dagger of F, given by

\[F^\dagger(y, x) = F(x, y).\]

dagger(d::OpenGaussianDistribution)