Library Reference

Gaussian Relations

AlgebraicInference.GaussianDistribution — Type

GaussianDistribution(Q, a)

A Gaussian distribution with covariance Q and mean a.

AlgebraicInference.GaussianDistribution — Method

GaussianDistribution(Q::AbstractMatrix)

Construct a centered Gaussian distribution with covariance Q.

AlgebraicInference.GaussianDistribution — Method

GaussianDistribution(a::AbstractVector)

Construct a Dirac distribution with mean a.

AlgebraicInference.GaussianRelation — Type

GaussianRelation(rf)

A Gaussian relation is a morphism in the hypergraph category $\text{Cond}(\text{GaussEx})$. Internally, it is represented as an equivalence class of partial quadratic bifunctions, where

\[F_1 \sim F_2 \iff F_1 = F_2 + \alpha \text{ for some } \alpha \in \mathbb{R}.\]

struct GaussianRelation{T₁, T₂, T₃, T₄}
    rf::QuadraticBifunction{T₁, T₂, T₃, T₄}
end

Every Gaussian relation $d: m \to n$ is either an extended Gaussian distribution on $\mathbb{R}^{m + n}$ or a failure state $\bot_{m + n}$.

References:

Stein & Staton (2021), "Compositional Semantics for Probabilistic Programs with Exact Conditioning" (arXiv:2101.11351)

Stein (2022), "Decorated Linear Relations: Extending Gaussian Probability with Uninformative Priors" (arXiv:2204.14024)

Stein (2022), "A Hypergraph Category for Exact Gaussian Inference", (https://msp.cis.strath.ac.uk/act2022/papers/ACT2022paper3601.pdf)

AlgebraicInference.GaussianRelation — Method

GaussianRelation(L::AbstractMatrix)

Construct the extended Gaussian distribution

\[\mathcal{N}(0, 0) + \{ (x, y) \mid Lx = y \}.\]

AlgebraicInference.GaussianRelation — Method

GaussianRelation(ψ::GaussianDistribution)

Construct the extended Gaussian distribution $\psi + \{ 0 \}$.

AlgebraicInference.GaussRelDom — Type

GaussRelDom(n)

The Euclidean space $\mathbb{R}^n$.

StatsAPI.params — Function

params(d::GaussianRelation)

Represent d as an extended Gaussian distribution.

Returns a quadruple $(Q, a, B, b)$. If $b \neq 0$, then $d = \bot$. Otherwise,

\[d = \mathcal{N}(a, Q) + \text{null } B.\]

Statistics.cov — Function

cov(ψ::GaussianDistribution)

Get the covariance of ψ.

cov(d::GaussianRelation)

Get the covariance of d.

Statistics.mean — Function

mean(ψ::GaussianDistribution)

Get the mean of ψ.

mean(d::GaussianRelation)

Get the mean of d.

Quadratic Functions

AlgebraicInference.QuadraticFunction — Type

QuadraticFunction(Q, a)

A quadratic function of the form

\[f(x) = \langle x, Qx + a \rangle,\]

where Q is positive semidefinite.

AlgebraicInference.QuadraticFunction — Method

QuadraticFunction(Q::AbstractMatrix)

Construct the quadratic function $f(x) = \langle x, Qx \rangle$, where $Q$ is positive semidefinite.

AlgebraicInference.QuadraticFunction — Method

QuadraticFunction(a::AbstractVector)

Construct the linear function $f(x) = \langle x, a \rangle$.

AlgebraicInference.QuadraticBifunction — Type

QuadraticBifunction(L, R, f)

A partial quadratic bifunction $F: m \to n$ is a convex function on $\mathbb{R}^{m + n}$ of the form

\[F(x, y) = \begin{cases} \langle (x, y), Q(x, y) + a \rangle + \alpha & B(x, y) = b \\ \infty & \text{else} \end{cases},\]

where $Q$ is positive semidefinite. Internally, $F$ is represented by a triple $(L, R, f)$ such that

\[F(x, y) = f^*(Ry - Lx),\]

where $f^*$ is the convex conjuate of $f$.

struct QuadraticBifunction{T₁ <: AbstractMatrix, T₂ <: AbstractMatrix, T₃, T₄}
    L::T₁
    R::T₂
    f::QuadraticFunction{T₃, T₄}
end

AlgebraicInference.QuadraticBifunction — Method

QuadraticBifunction(L::AbstractMatrix)

Construct the indicator bifunction corresponding to L, given by

\[F(x, y) = \begin{cases} 0 & Lx = y \\ \infty & \text{else} \end{cases}.\]

AlgebraicInference.QuadraticBifunction — Method

QuadraticBifunction(f::QuadraticFunction)

Construct the bifunction $F(*, y) = f^*(y)$, where $f^*$ is the convex conjugate of $f$.

AlgebraicInference.QuadDom — Type

QuadDom(n)

The Euclidian space $\mathbb{R}^n$.

Base.:* — Function

*(f::QuadraticFunction, M::Union{UniformScaling, AbstractMatrix})

Compute the inverse image of f under M, given by

\[(fM)(y) = f(My).\]

*(M::Union{UniformScaling, AbstractMatrix}, f::QuadraticFunction)

Compute the image of f under M, given by

\[(Mf)(y) = \inf \{ f(x) \mid y = Mx \}.\]

Returns a quintuple $(Q, a, \alpha, B, b).$ If $b \neq 0$, then $Mf = -\infty$. Otherwise,

\[(Mf)(y) = \begin{cases} \langle y, \frac{1}{2}Qy + a \rangle + \alpha & By = 0 \\ \infty & \text{else} \end{cases}.\]

AlgebraicInference.conjugate — Function

conjugate(F::QuadraticBifunction)

Compute the convex conjugate of F, given by

\[F^*(x^*, y^*) = \sup \{ \langle y, y^* \rangle - \langle x, x^* \rangle - F(x, y) \mid x, y \}.\]

Returns a quintuple $(Q, a, \alpha, B, b).$ If $b \neq 0$, then $F^* = -\infty$. Otherwise,

\[F^*(x^*, y^*) = \begin{cases} \langle (x^*, y^*), \frac{1}{2}Q(x^*, y^*) + a \rangle + \alpha & By^* = 0 \\ \infty & \text{else} \end{cases}.\]

Catlab.Schemas.dom — Function

dom(F::QuadraticBifunction)

The domain of a bifunction is the dimension of its first variable.

Catlab.Schemas.codom — Function

codom(F::QuadraticBifunction)

The codomain of a bifunction is the dimension of its second variable.

Catlab.Theories.mzero — Function

mzero(::Type{QuadDom})

The Euclidean space $\mathbb{R}^0$.

Catlab.Theories.dagger — Function

dagger(F::QuadraticBifunction)

Compute the dagger of F, given by

\[F^\dagger(y, x) = F(x, y).\]

Catlab.Theories.compose — Function

compose(F₁::QuadraticBifunction, F₂::QuadraticFunction)

Compute the product of F₁ and F₂, given by

\[ (F_2F_1)(x, z) = \sup \{ F_1(x, y) + F_2(y, z) \mid y \}.\]

Catlab.Theories.oplus — Function

oplus(f₁::QuadraticFunction, f₂::QuadraticFunction)

Compute the direct sum of f₁ and f₂, given by

\[(f_1 \oplus f_2)(x, y) = f_1(x) + f_2(y).\]

oplus(X::QuadDom, Y::QuadDom)

Compute the direct sum of $X$ and $Y$.

oplus(F₁::QuadraticBifunction, F₂::QuadraticBifunction)

Compute the direct sum of F₁ and F₂, given by

\[ (F_1 \oplus F_2)((x_1, x_2), (y_1, y_2)) = F_1(x_1, y_1) + F_2(x_2, y_2).\]

Catlab.Theories.meet — Function

meet(F₁::QuadraticBifunction, F₂::QuadraticBifunction)

Compute the sum of F₁ and F₂, given by

\[ (F_1 + F_2)(x, y) = F_1(x, y) + F_2(x, y).\]

Base.join — Function

join(F₁::QuadraticBifunction, F₂::QuadraticBifunction)

Compute the infimal convolution of F₁ and F₂, given by

\[ (F_1 \Box F_2)(x^*, y^*) = \inf \{ F_1(x^*, y^*) + F_2(x - x^*, y - y^*) \mid x, y \}.\]

Catlab.Theories.id — Function

id(X::QuadDom)

Construct the indicator bifunction

\[F(x, y) = \begin{cases} 0 & x = y \\ \infty & \text{else} \end{cases}.\]

Base.zero — Function

zero(X::QuadDom)

Construct the indicator bifunction $F(*, y) = 0.$

Catlab.Theories.delete — Function

delete(X::QuadDom)

Construct the indicator bifunction $F(x, *) = 0.$

Catlab.Theories.mcopy — Function

mcopy(X::QuadDom)

Construct the indicator bifunction

\[F(x, (y_1, y_2)) = \begin{cases} 0 & x = y_1 = y_2 \\ \infty & \text{else} \end{cases}.\]

Catlab.Theories.plus — Function

plus(X::QuadDom)

Construct the indicator bifunction

\[F((x_1, x_2), y) = \begin{cases} 0 & x_1 + x_2 = y \\ \infty & \text{else} \end{cases}.\]

Catlab.Theories.dunit — Function

dunit(X::QuadDom)

Construct the indicator bifunction

\[F(*, (y_1, y_2)) = \begin{cases} 0 & y_1 = y_2 \\ \infty & \text{else} \end{cases}.\]

Catlab.Theories.cozero — Function

cozero(X::QuadDom)

Construct the indicator bifunction

\[F(x, *) = \begin{cases} 0 & x = 0 \\ \infty & \text{else} \end{cases}.\]

Catlab.Theories.create — Function

create(X::QuadDom)

Construct the indicator bifunction

\[F(*, y) = \begin{cases} 0 & y = 0 \\ \infty & \text{else} \end{cases}\]

Catlab.Theories.mmerge — Function

mmerge(X::QuadDom)

Construct the indicator bifunction

\[F((x_1, x_2), y) = \begin{cases} 0 & x_1 = x_2 = y \\ \infty & \text{else} \end{cases}.\]

Catlab.Theories.coplus — Function

coplus(X::QuadDom)

Construct the indicator bifunction

\[F((x_1, x_2), y) = \begin{cases} 0 & x_1 + x_2 = y \\ \infty & \text{else} \end{cases}\]

Catlab.Theories.dcounit — Function

dcounit(X::QuadDom)

Construct the indicator bifunction

\[F((x_1, x_2), *) = \begin{cases} 0 & x_1 = x_2 \\ \infty & \text{else} \end{cases}.\]

Catlab.Theories.swap — Function

swap(X::QuadDom, Y::QuadDom)

Construct the indicator bifunction

\[F((x_1, x_2), (y_1, y_2)) = \begin{cases} 0 & x_1 = y_2, x_2 = y_1 \\ \infty & \text{else} \end{cases}.\]

Catlab.Theories.top — Function

top(X::QuadDom, Y::QuadDom)

Construct the indicator bifunction

\[F(x, y) = \begin{cases} 0 & x = y = 0 \\ \infty & \text{else} \end{cases}.\]

Catlab.Theories.bottom — Function

bottom(X::QuadDom, Y::QuadDom)

Construct the indicator bifunction $F(x, y) = 0.$