e# Library Reference

Systems

AlgebraicInference.GaussianSystem — Type

GaussianSystem{T₁, T₂, T₃, T₄, T₅}

A Gaussian system.

AlgebraicInference.CanonicalForm — Type

CanonicalForm{T₁, T₂} = GaussianSystem{
    T₁,
    ZerosMatrix{Bool, Tuple{OneTo{Int}, OneTo{Int}}},
    T₂,
    ZerosVector{Bool, Tuple{OneTo{Int}}},
    Bool}

A canonical form.

AlgebraicInference.DenseGaussianSystem — Type

DenseGaussianSystem{T} = GaussianSystem{
    Matrix{T},
    Matrix{T},
    Vector{T},
    Vector{T},
    T}

A Gaussian system represented by dense matrices and vectors.

AlgebraicInference.DenseCanonicalForm — Type

DenseCanonicalForm{T} = CanonicalForm{
    Matrix{T},
    Vector{T}}

A canonical form represented by a dense matrix and vector.

AlgebraicInference.GaussianSystem — Method

GaussianSystem(
    P::AbstractMatrix,
    S::AbstractMatrix,
    p::AbstractVector,
    s::AbstractVector,
    σ::Real)

Construct a Gaussian system by specifying its energy function. You should set

σ  = dot(s, pinv(S) * s)

AlgebraicInference.CanonicalForm — Method

CanonicalForm(K::AbstractMatrix, h::AbstractVector)

Construct the canonical form

\[\mathcal{C}(K, h, g),\]

where the normalization constant $g$ is inferred from $K$ and $h$.

AlgebraicInference.normal — Function

normal(μ::AbstractVector, Σ::AbstractMatrix)

Construct a multivariate normal distribution with mean vector μ and covariance matrix Σ.

normal(μ::Real, σ::Real)

Construct a normal distribution with mean μ and standard deviation σ.

AlgebraicInference.kernel — Function

kernel(L::AbstractMatrix, μ::AbstractVector, Σ::AbstractMatrix)

Construct a conditional distribution of the form

\[p(Y \mid X = x) = \mathcal{N}(Lx + \mu, \Sigma).\]

kernel(l::AbstractVector, μ::Real, σ::Real)

Construct a conditional distribution of the form

\[p(Y \mid X = x) = \mathcal{N}(l^\mathsf{T}x + \mu, \sigma^2).\]

Base.length — Method

length(Σ::GaussianSystem)

Get the dimension of Σ.

Statistics.cov — Method

cov(Σ::GaussianSystem; atol::Real=1e-8)

Get the covariance matrix of Σ.

Distributions.invcov — Method

invcov(Σ::GaussianSystem)

Get the precision matrix of Σ.

Statistics.var — Method

var(Σ::GaussianSystem; atol::Real=1e-8)

Get the variances of Σ.

Statistics.mean — Method

mean(Σ::GaussianSystem; atol::Real=1e-8)

Get the mean vector of Σ.

Problems

AlgebraicInference.InferenceProblem — Type

InferenceProblem{T₁, T₂, T₃, T₄}

An inference problem.

AlgebraicInference.InferenceProblem — Method

InferenceProblem(
    diagram::RelationDiagram,
    hom_map::AbstractDict,
    ob_map::AbstractDict;
    hom_attr::Symbol=:name,
    ob_attr::Symbol=:junction_type,
    var_attr::Symbol=:variable
    check::Bool=true)

Construct an inference problem that performs undirected compositon.

AlgebraicInference.InferenceProblem — Method

InferenceProblem(
    network::BayesNet,
    query::AbstractVector,
    context::AbstractDict)

Construct an inference problem that queries a Bayesian network.

CommonSolve.solve — Method

solve(
    problem::InferenceProblem,
    elimination_algorithm::EliminationAlgorithm=MinFill()
    supernode_type::SupernodeType=Node()
    architecture_type::ArchitectureType=ShenoyShafer())

Solve an inference problem. This involves several steps:

interaction graph
    ↓ elimination algorithm
elimination tree
    ↓ supernode type
join tree
    ↓ architecture type
solution

First an elimination agorithm is used to construct an elimination tree from the model's interaction graph. Then, nodes of the elimination tree are merged into supernodes, forming a join (or junction) tree. Finally, a message passing algorithm is used to compute a solution.

CommonSolve.init — Method

init(
    problem::InferenceProblem,
    elimination_algorithm::EliminationAlgorithm=MinFill(),
    supernode_type::SupernodeType=Node(),
    architecture_type::ArchitectureType=ShenoyShafer())

Construct a solver for an inference problem. This involves several steps:

interaction graph
    ↓ elimination algorithm
elimination tree
    ↓ supernode type
join tree

First an elimination agorithm is used to construct an elimination tree from the model's interaction graph. Then, nodes of the elimination tree are merged into supernodes, forming a join (or junction) tree.

When solve! is called on the solver, a message passing algorithm is used to compute the solution.

Solvers

AlgebraicInference.InferenceSolver — Type

InferenceSolver{T₁, T₂, T₃, T₄, T₅}

A solver for an inference problem.

CommonSolve.solve! — Method

solve!(solver::InferenceSolver)

Solve an inference problem.

Statistics.mean — Method

mean(solver::InferenceSolver{AncestralSampler()})

Base.rand — Method

rand(solver::InferenceSolver{AncestralSampler()})

Base.rand — Method

rand(rng::AbstractRNG, solver::InferenceSolver{AncestralSampler()})

Elimination

AlgebraicInference.EliminationAlgorithm — Type

EliminationAlgorithm

A graph elimination algorithm. The options are

MinDegree

MinFill

MaxCardinality

ChordalGraph

CuthillMcKeeJL_RCM

AMDJL_AMD

MetisJL_ND

AlgebraicInference.MaxCardinality — Type

MaxCardinality <: EliminationAlgorithm

The maximum cardinality search algorithm.

AlgebraicInference.MinDegree — Type

MinDegree <: EliminationAlgorithm

The minimum-degree heuristic.

AlgebraicInference.MinFill — Type

MinFill <: EliminationAlgorithm

The minimum-fill heuristic.

AlgebraicInference.ChordalGraph — Type

ChordalGraph <: EliminationAlgorithm

An efficient algorithm for chordal graphs.

AlgebraicInference.CuthillMcKeeJL_RCM — Type

CuthillMcKeeJL_RCM <: EliminationAlgorithm

The reverse Cuthill-McKee algorithm. Uses CuthillMckee.jl.

AlgebraicInference.AMDJL_AMD — Type

AMDJL_AMD <: EliminationAlgorithm

The approximate minimum degree algorithm. Uses AMD.jl.

AlgebraicInference.MetisJL_ND — Type

MetisJL_ND <: EliminationAlgorithm

The nested dissection heuristic. Uses Metis.jl.

Supernodes

AlgebraicInference.SupernodeType — Type

SupernodeType

A type of supernode. The options are

Node

MaximalSupernode

FundamentalSupernode

AlgebraicInference.Node — Type

Node <: SupernodeType

The single-vertex supernode.

AlgebraicInference.MaximalSupernode — Type

MaximalSupernode <: SupernodeType

The maximal supernode.

AlgebraicInference.FundamentalSupernode — Type

FundamentalSupernode <: SupernodeType

The fundamental supernode.

Architectures

AlgebraicInference.ArchitectureType — Type

ArchitectureType

A message-passing algorithm. The options are

ShenoyShafer

LauritzenSpiegelhalter

HUGIN

Idempotent

There is one additional type, which you can use for sampling from a graphical model:

AncestralSampler

AlgebraicInference.ShenoyShafer — Type

ShenoyShafer <: ArchitectureType

The Shenoy-Shafer architecture.

AlgebraicInference.LauritzenSpiegelhalter — Type

LauritzenSpiegelhalter <: ArchitectureType

The Lauritzen-Spiegelhalter architecture.

AlgebraicInference.HUGIN — Type

HUGIN <: ArchitectureType

The HUGIN architecture.

AlgebraicInference.Idempotent — Type

Idempotent <: ArchitectureType

The idempotent architecture.

AlgebraicInference.AncestralSampler — Type

AncestralSampler <: ArchitectureType

A type for sampling graphical models.