Library Reference

Gaussian Relations

AlgebraicInference.GaussianDistribution — Type

GaussianDistribution(Q, a)

A Gaussian distribution with covariance matrix $Q$ and mean vector $a$.

AlgebraicInference.GaussianDistribution — Method

GaussianDistribution(Q::AbstractMatrix)

Construct a centered Gaussian distribution with covariance matrix $Q$.

AlgebraicInference.GaussianDistribution — Method

GaussianDistribution(a::AbstractVector)

Construct a Dirac distribution with mean vector $a$.

AlgebraicInference.GaussianRelation — Type

GaussianRelation(logdensity)

The Gaussian relation determined by quadratic bifunction logdensity.

A Gaussian relation is a equivalence class of quadratic bifunctions. Bifunctions $F_1$ and $F_2$ are equivalent if $F_1 = F_2 + \alpha$ for some $\alpha \in \mathbb{R}$.

If $F: m \to n$ and $F \neq -\infty$, then the Gaussian relation determined by $F$ corresponds to an extended Gaussian distribution on $\mathbb{R}^{m + n}$.

AlgebraicInference.GaussianRelation — Method

GaussianRelation(L::AbstractMatrix)

Construct the extended Gaussian distribution

\[\mathcal{N}(0, 0) + \{ (x, y) \mid Lx = y \}.\]

AlgebraicInference.GaussianRelation — Method

GaussianRelation(ψ::GaussianDistribution)

Construct the extended Gaussian distribution $\psi + \{ 0 \}$.

AlgebraicInference.GaussRelDom — Type

GaussRelDom(n)

The Euclidean space $\mathbb{R}^m$.

StatsAPI.params — Function

params(d::GaussianRelation)

Represent a Gaussian relation as an extended Gaussian distribution.

Returns a quadruple $(Q, a, B, b)$. If $b = 0$, then $d$ corresponds to the extended Gaussian distribution

\[ \mathcal{N}(a, Q) + \text{null } B.\]

Otherwise, $d$ does not correspond to an extended Gaussian distribution.

Statistics.cov — Function

cov(ψ::GaussianDistribution)

Return the covariance matrix associated to $\Sigma$.

cov(d::GaussianRelation)

Return a version of the covariance matrix associated to $d$.

Statistics.mean — Function

mean(ψ::GaussianDistribution)

Return the mean vector associated to $\psi$.

mean(d::GaussianRelation)

Return a version of the mean vector associated to $d$.

Quadratic Functions

AlgebraicInference.QuadraticFunction — Type

QuadraticFunction(Q, a)

A convex quadratic function $f$ of the form

\[f(x) = \langle x, Qx + a \rangle,\]

where $Q$ is positive semidefinite.

AlgebraicInference.QuadraticFunction — Method

QuadraticFunction(Q::AbstractMatrix)

Construct the convex quadratic function $f(x) = \langle x, Qx \rangle$, where $Q$ is positive semidefinite.

AlgebraicInference.QuadraticFunction — Method

QuadraticFunction(a::AbstractVector)

Construct the linear function $f(x) = \langle x, a \rangle$.

AlgebraicInference.QuadraticBifunction — Type

QuadraticBifunction(L, R, f)

A concave quadratic bifunction $F$ of the form

\[F(x, y) = -f^*(Lx - Ry),\]

where $f^*$ is the convex conjuate of $f$.

AlgebraicInference.QuadraticBifunction — Method

QuadraticBifunction(L::AbstractMatrix)

Construct the concave indicator bifunction corresponding to $L$, given by

\[F(x, y) = \begin{cases} 0 & Lx = y \\ -\infty & \text{else} \end{cases}.\]

AlgebraicInference.QuadraticBifunction — Method

QuadraticBifunction(f::QuadraticFunction)

Construct the concave bifunction $F(*, y) = -f^*(-y)$, where $f^*$ is the convex conjugate of $f$.

AlgebraicInference.QuadDom — Type

QuadDom(n)

The Euclidian space $\mathbb{R}^n$.

Base.:* — Function

*(f::QuadraticFunction, M::Union{UniformScaling, AbstractMatrix})

Compute the inverse image of $f$ under $M$, given by

\[(fM)(y) = f(My).\]

*(M::Union{UniformScaling, AbstractMatrix}, f::QuadraticFunction)

Compute the image of $f$ under $M$, given by

\[(Mf)(y) = \inf \{ f(x) \mid y = Mx \}.\]

Returns a quintuple $(Q, a, \alpha, B, b).$ If $b \neq 0$, then $Mf = -\infty$. Otherwise,

\[(Mf)(y) = \begin{cases} \langle y, \frac{1}{2}Qy + a \rangle + \alpha & By = 0 \\ \infty & \text{else} \end{cases}.\]

Base.adjoint — Function

adjoint(F::QuadraticBifunction)

Compute the adjoint of $F$, given by

\[F^*(y^*, x^*) = \sup \{ F(x, y) - \langle y, y^* \rangle + \langle x, x^* \rangle \mid x, y \}.\]

Returns a quintuple $(Q, a, \alpha, B, b).$ If $b \neq 0$, then $F^* = -\infty$. Otherwise,

\[(F^*)(y, x) = \begin{cases} \langle (y, x), \frac{1}{2}Q(y, x) + a \rangle + \alpha & By = 0 \\ \infty & \text{else} \end{cases}.\]

Catlab.Schemas.dom — Function

dom(F::QuadraticBifunction)

The domain of a bifunction is the dimension of its first variable.

Catlab.Schemas.codom — Function

codom(F::QuadraticBifunction)

The codomain of a bifunction is the dimension of its second variable.

Catlab.Theories.mzero — Function

mzero(::Type{QuadDom})

The Euclidean space $\mathbb{R}^0$.

Catlab.Theories.dagger — Function

dagger(F::QuadraticBifunction)

Compute the dagger of $F$, given by

\[F^\dagger(y, x) = F(x, y).\]

Catlab.Theories.compose — Function

compose(F₁::QuadraticBifunction, F₂::QuadraticFunction)

Compute the product of $F_1$ and $F_2$, given by

\[ (F_2F_1)(x, z) = \sup \{ F_1(x, y) + F_2(y, z) \mid y \}.\]

Catlab.Theories.oplus — Function

oplus(f₁::QuadraticFunction, f₂::QuadraticFunction)

Compute the direct sum of $f_1$ and $f_2$, given by

\[(f_1 \oplus f_2)(x, y) = f_1(x) + f_2(y).\]

oplus(X::QuadDom, Y::QuadDom)

Compute the direct sum of $X$ and $Y$.

oplus(F₁::QuadraticBifunction, F₂::QuadraticBifunction)

Compute the direct sum of $F_1$ and $F_2$, given by

\[ (F_1 \oplus F_2)((x_1, x_2), (y_1, y_2)) = F_1(x_1, y_1) + F_2(x_2, y_2).\]

Catlab.Theories.meet — Function

meet(F₁::QuadraticBifunction, F₂::QuadraticBifunction)

Compute the sum of $F_1$ and $F_2$, given by

\[ (F_1 + F_2)(x, y) = F_1(x, y) + F_2(x, y).\]

Base.join — Function

join(F₁::QuadraticBifunction, F₂::QuadraticBifunction)

Compute the supremal convolution of $F_1$ and $F_2$, given by

\[ (F_1 \Box F_2)(x, y) = \sup \{ F_1(x_1, y_1) + F_2(x - x_1, y - y_1) \mid x, y \}.\]

Catlab.Theories.id — Function

id(X::QuadDom)

Construct the concave indicator bifunction

\[F(x, y) = \begin{cases} 0 & x = y \\ -\infty & \text{else} \end{cases}.\]

Base.zero — Function

zero(X::QuadDom)

Construct the constant bifunction $F(*, y) = 0.$

Catlab.Theories.delete — Function

delete(X::QuadDom)

Construct the constant bifunction $F(x, *) = 0.$

Catlab.Theories.mcopy — Function

mcopy(X::QuadDom)

Construct the concave indicator bifunction

\[F(x, (y_1, y_2)) = \begin{cases} 0 & x = y_1 = y_2 \\ -\infty & \text{else} \end{cases}.\]

Catlab.Theories.plus — Function

plus(X::QuadDom)

Construct the concave indicator bifunction

\[F((x_1, x_2), y) = \begin{cases} 0 & x_1 + x_2 = y \\ -\infty & \text{else} \end{cases}.\]

Catlab.Theories.dunit — Function

dunit(X::QuadDom)

Construct the concave indicator bifunction

\[F(*, (y_1, y_2)) = \begin{cases} 0 & y_1 = y_2 \\ -\infty & \text{else} \end{cases}.\]

Catlab.Theories.cozero — Function

cozero(X::QuadDom)

Construct the concave indicator bifunction

\[F(x, *) = \begin{cases} 0 & x = 0 \\ -\infty & \text{else} \end{cases}.\]

Catlab.Theories.create — Function

create(X::QuadDom)

Construct the concave indicator bifunction

\[F(*, y) = \begin{cases} 0 & y = 0 \\ -\infty & \text{else} \end{cases}\]

Catlab.Theories.mmerge — Function

mmerge(X::QuadDom)

Construct the concave indicator bifunction

\[F((x_1, x_2), y) = \begin{cases} 0 & x_1 = x_2 = y \\ -\infty & \text{else} \end{cases}.\]

Catlab.Theories.coplus — Function

coplus(X::QuadDom)

Construct the concave indicator bifunction

\[F((x_1, x_2), y) = \begin{cases} 0 & x_1 + x_2 = y \\ -\infty & \text{else} \end{cases}\]

Catlab.Theories.dcounit — Function

dcounit(X::QuadDom)

Construct the concave indicator bifunction

\[F((x_1, x_2), *) = \begin{cases} 0 & x_1 = x_2 \\ -\infty & \text{else} \end{cases}.\]

Catlab.Theories.swap — Function

swap(X::QuadDom, Y::QuadDom)

Construct the concave indicator bifunction

\[F((x_1, x_2), (y_1, y_2)) = \begin{cases} 0 & x_1 = y_2, x_2 = y_1 \\ -\infty & \text{else} \end{cases}.\]

Catlab.Theories.top — Function

top(X::QuadDom, Y::QuadDom)

Construct the concave indicator bifunction

\[F(x, y) = \begin{cases} 0 & x = y = 0 \\ -\infty & \text{else} \end{cases}.\]

Catlab.Theories.bottom — Function

bottom(X::QuadDom, Y::QuadDom)

Construct the constant bifunction $F(x, y) = 0.$