Library Reference

Systems

AlgebraicInference.GaussianSystem — Type

GaussianSystem{
    T₁ <: AbstractMatrix,
    T₂ <: AbstractMatrix, 
    T₃ <: AbstractVector,
    T₄ <: AbstractVector,
    T₅}

A Gaussian system.

AlgebraicInference.GaussianSystem — Method

GaussianSystem(
    P::AbstractMatrix,
    S::AbstractMatrix,
    p::AbstractVector,
    s::AbstractVector,
    σ)

Construct a Gaussian system by specifying its energy function.

Set $\sigma = s^\mathsf{T} S^+ s$, where $S^+$ is the Moore-Penrose psuedoinverse of $S$.

AlgebraicInference.canon — Method

canon(J::AbstractMatrix, h::AbstractVector)

Construct a multivariate normal distribution with information matrix J and potential vector h.

AlgebraicInference.canon — Method

canon(J::AbstractMatrix)

Construct a centered multivariate normal distribution with information matrix J.

AlgebraicInference.normal — Method

normal(Σ::AbstractMatrix, μ::AbstractVector)

Construct a multivariate normal distribution with covariance matrix Σ and mean vector μ.

AlgebraicInference.normal — Method

normal(Σ::AbstractMatrix)

Construct a centered multivariate normal distribution with covariance matrix Σ.

AlgebraicInference.normal — Method

normal(μ::AbstractVector)

Construct a Dirac distribution with mean vector μ.

AlgebraicInference.kernel — Method

kernel(Σ::AbstractMatrix, μ::AbstractVector, L::AbstractMatrix)

Construct a conditional distribution of the form

\[ y \mid x \sim \mathcal{N}(Lx + \mu, \Sigma).\]

AlgebraicInference.kernel — Method

kernel(Σ::AbstractMatrix, L::AbstractMatrix)

Construct a conditional distribution of the form

\[ y \mid x \sim \mathcal{N}(Lx, \Sigma).\]

AlgebraicInference.kernel — Method

kernel(L::AbstractMatrix)

Construct a conditional distribution of the form

\[ y \mid x \sim \delta_{Lx}.\]

Base.length — Method

length(Σ::GaussianSystem)

Get the dimension of Σ.

Statistics.cov — Method

cov(Σ::GaussianSystem)

Get the covariance matrix of Σ.

AlgebraicInference.invcov — Method

invcov(Σ::GaussianSystem)

Get the information matrix of Σ.

Statistics.mean — Method

mean(Σ::GaussianSystem)

Get the mean vector of Σ.

Catlab.Theories.:⊗ — Method

⊗(Σ₁::GaussianSystem, Σ₂::GaussianSystem)

Compute the tensor product of Σ₁ and Σ₂.

Base.:+ — Method

+(Σ₁::GaussianSystem, Σ₂::GaussianSystem)

Construct a Gaussian system by summing the energy functions of Σ₁ and Σ₂.

Base.:* — Method

*(Σ::GaussianSystem, M::AbstractMatrix)

Construct a Gaussian system with energy function $E'(x) = E(Mx),$ where $E$ is the energy function of Σ.

Base.zero — Method

zero(Σ::GaussianSystem)

Construct a Gaussian system with energy function $E(x) = 0$.

AlgebraicInference.pushfwd — Function

pushfwd(M::AbstractMatrix, Σ::GaussianSystem)

Compute the pushforward $M_*\Sigma$.

AlgebraicInference.marginal — Function

marginal(m::AbstractVector{Bool}, Σ::GaussianSystem)

Compute the marginal of Σ along the indices specified by m.

Catlab.WiringDiagrams.WiringDiagramAlgebras.oapply — Method

oapply(wd::UndirectedWiringDiagram, box_map::AbstractDict{<:Any, <:GaussianSystem})

See oapply(wd::UndirectedWiringDiagram, boxes::AbstractVector{<:GaussianSystem}).

Catlab.WiringDiagrams.WiringDiagramAlgebras.oapply — Method

oapply(wd::UndirectedWiringDiagram, boxes::AbstractVector{<:GaussianSystem})

Compose Gaussian systems according to the undirected wiring diagram wd.

Valuations

AlgebraicInference.Valuation — Type

Valuation{T}

Abstract type for valuations in a valuation algebra.

Subtypes should specialize the following methods:

domain(ϕ::Valuation)

combine(ϕ₁::Valuation, ϕ₂::Valuation)

project(ϕ::Valuation, x)

Valuations are parametrized by the type of the variables in their variable system. If isa(ϕ, Valuation{T}), then domain(ϕ) should return a container with element type T.

AlgebraicInference.IdentityValuation — Type

IdentityValuation{T} <: Valuation{T}

The identity element $e$.

AlgebraicInference.LabeledBox — Type

LabeledBox{T₁, T₂} <: Valuation{T₁}

A filler for a box in an undirected wiring diagram, labeled with the junctions to which the box is incident.

AlgebraicInference.domain — Function

domain(ϕ::Valuation)

Get the domain of $\phi$.

AlgebraicInference.combine — Function

combine(ϕ₁::Valuation, ϕ₂::Valuation)

Perform the combination $\phi_1 \otimes \phi_2$.

AlgebraicInference.project — Function

project(ϕ::Valuation, x)

Perform the projection $\phi^{\downarrow x}$.

AlgebraicInference.inference_problem — Method

inference_problem(wd::UndirectedWiringDiagram, box_map::AbstractDict)

See inference_problem(wd::UndirectedWiringDiagram, boxes::AbstractVector).

AlgebraicInference.inference_problem — Method

inference_problem(wd::UndirectedWiringDiagram, boxes::AbstractVector)

Translate an undirected wiring diagram

\[ B \xleftarrow{\mathtt{box}} P \xrightarrow{\mathtt{junc}} J \xleftarrow{\mathtt{junc'}} Q\]

into an inference problem in a valuation algebra.

The diagram must satisfy the following constraints:

$\mathtt{junc'}$ is injective.

$\mathtt{image}(\mathtt{junc'}) \subseteq \mathtt{image}(\mathtt{junc})$

For all $x, y \in P$, $\mathtt{box}(x) = \mathtt{box}(y)$ and $\mathtt{junc}(x) = \mathtt{junc}(y)$ implies that $x = y$.

Architectures

AlgebraicInference.Architecture — Type

Architecture{T₁, T₂} <: AbstractNode{T₁}

A join tree $(V, E, \lambda, D)$, along with a set of factors $\left\{ \phi_i \right\}_{i \in V}$ and mailboxes

\[ \left\{ \left( \mu_{ i \to \mathtt{pa}(i)}, \mu_{\mathtt{pa}(i) \to i} \right) \right\}_{i \in V}.\]

AlgebraicInference.architecture — Function

architecture(kb::AbstractVector{<:Valuation{T}}, order) where T

Construct a covering join tree for the knowledge base kb using the variable elimination order order.

AlgebraicInference.answer_query — Function

answer_query(jt::Architecture, query)

Answer a query.

AlgebraicInference.answer_query! — Function

answer_query!(jt::Architecture, query)

Answer a query, caching intermediate computations in the mailboxes of jt.

Graphs

AlgebraicInference.primal_graph — Function

primal_graph(kb::AbstractVector{<:Valuation{T}}) where T

Construct the primal graph of the knowledge base kb.

AlgebraicInference.minfill! — Function

minfill!(g::MetaGraph, query)

Compute a vertex elimination order using the min-fill heuristic.

AlgebraicInference.minwidth! — Function

minwidth!(g::MetaGraph, query)

Compute a vertex elimination order using the min-width heuristic