Library Reference

Architectures

AlgebraicInference.Architecture — Type

Architecture{T₁, T₂} <: AbstractNode{T₁}

An Architecture contains a join tree $(V, E, \lambda, D)$. To each vertex $i \in V$ it optionally associates a factor $\psi_i$ and mailbox

\[\left( \mu_{i \to \mathtt{pa}(i)}, \mu_{\mathtt{pa}(i) \to i} \right).\]

AlgebraicInference.architecture — Function

architecture(kb::AbstractVector{<:Valuation{T}}, order) where T

Construct a covering join tree for the knowledge base kb using a variable elimination order order.

AlgebraicInference.answer_query — Function

answer_query(jt::Architecture{T₁, T₂}, query::Set{T₂}) where {T₁, T₂}

Answer a query.

Let $(V, E, \lambda, D)$ be a join tree with factors $\{\phi_i\}_{i \in V}$ and $x$ a query covered by $(V, E, \lambda, D)$. Then answer_query(jt, query) solves the inference problem

\[\left( \bigotimes_{i \in V} \psi_i \right)^{\downarrow x}.\]

AlgebraicInference.answer_query! — Function

answer_query!(jt::Architecture{T₁, T₂}, query::Set{T₂}) where {T₁, T₂}

Answer a query, caching intermediate computations.

Let $(V, E, \lambda, D)$ be a join tree with factors $\{\phi_i\}_{i \in V}$ and $x$ be a query covered by $(V, E, \lambda, D)$. Then answer_query!(jt, query) solves the inference problem

\[\left( \bigotimes_{i \in V} \psi_i \right)^{\downarrow x}.\]

Graphs

AlgebraicInference.primal_graph — Function

primal_graph(kb::AbstractVector{<:Valuation{T}}) where T

Construct the primal graph of the knowledge base kb.

AlgebraicInference.minfill! — Function

minfill!(g::MetaGraph)

Compute a vertex elimination order using the min-fill heuristic.

AlgebraicInference.minwidth! — Function

minwidth!(g::MetaGraph)

Compute a vertex elimination order using the min-width heuristic

Systems

AlgebraicInference.AbstractSystem — Type

AbstractSystem

Abstract type for Gaussian systems.

Subtypes should specialize the following methods:

fiber(Σ::AbstractSystem)

mean(Σ::AbstractSystem)

cov(Σ::AbstractSystem)

*(M::AbstractMatrix, Σ::AbstractSystem)

\(M::AbstractMatrix, Σ::AbstractSystem)

⊗(Σ₁::AbstractSystem, Σ₂::AbstractSystem)

References:

J. C. Willems, "Open Stochastic Systems," in IEEE Transactions on Automatic Control, vol. 58, no. 2, pp. 406-421, Feb. 2013, doi: 10.1109/TAC.2012.2210836.

AlgebraicInference.AbstractProgram — Type

AbstractProgram <: AbstactSystem

A program in the Gaussian probabilistic programming language. Every program

\[ x: R^m \vdash e: R^n\]

can be interepreted as a Gaussian system $(\mathbb{R}^{m+n}, \mathcal{E}, P)$.

References

Stein, Dario and Sam Staton. "Compositional Semantics for Probabilistic Programs with Exact Conditioning." 2021 36th Annual ACM/IEEE Symposium on Logic in Computer Science (LICS) (2021): 1-13.

AlgebraicInference.ClosedProgram — Type

ClosedProgram{T₁ <: AbstractMatrix, T₂ <: AbstractVector} <: AbstractProgram

A closed program $\vdash e: R^n$.

AlgebraicInference.OpenProgram — Type

OpenProgram{T₁, T₂, T₂ <: AbstractMatrix} <: AbstractProgram

An open program $x: R^m \vdash e: R^n$.

AlgebraicInference.System — Type

System{T₁ <: AbstractMatrix, T₂, T₃} <: AbstractSystem

A Gaussian system. See AbstractSystem.

AlgebraicInference.ClosedProgram — Method

ClosedProgram(Γ::AbstractMatrix, μ::AbstractVector)

Construct a multivariate normal distribution with mean $\mu$ and covariance $\Gamma$.

AlgebraicInference.ClosedProgram — Method

ClosedProgram(Γ::AbstractMatrix)

Construct a centered multivariate normal distribution with covariance $\Gamma$.

AlgebraicInference.ClosedProgram — Method

ClosedProgram(μ::AbstractVector)

Construct a Dirac distribution with mean $\mu$.

AlgebraicInference.OpenProgram — Method

OpenProgram(ϵ::ClosedProgram, L::AbstractMatrix, o::Int)

Let $\vdash \epsilon : R^n * R^o$ be a closed program and $L$ an $(n + o) \times m$ matrix. Then OpenProgram(ϵ, L, o) constructs the open program

\[ x: R^m \vdash \text{let } (y, k): R^n * R^o = Lx + \epsilon \text{ in } (k := \mathbb{0}); y.\]

AlgebraicInference.OpenProgram — Method

OpenProgram(ϵ::ClosedProgram, L::AbstractMatrix)

Let $\vdash \epsilon : R^n$ be a closed program and $L$ an $m \times n$ matrix. Then OpenProgram(ϵ, L) constructs the open program

\[ x: R^m \vdash Lx + \epsilon.\]

In particular, if $\epsilon = \mathcal{N}(\mu, \Gamma)$, then OpenProgram(ϵ, L) constructs the conditional probability distribution

\[ (y \mid x) \sim \mathcal{N}(\mu + Lx, \Gamma).\]

AlgebraicInference.OpenProgram — Method

OpenProgram(L::AbstractMatrix)

AlgebraicInference.System — Method

System(ϵ::ClosedProgram, R::AbstractMatrix)

Let $R$ be an $m \times n$ matrix and $\epsilon$ an $m$-variate random vector with mean $\mu$ and covariance $\Gamma$.

If $\mu \in \mathtt{image}(R : \Gamma)$, then there exists a random variable $\hat{w}$ taking values in $(\mathbb{R}^n, \sigma R)$ that almost-surely solves the convex program

\[ \begin{align*} \underset{w}{\text{minimize }} & E(\epsilon, w) \\ \text{subject to } & Rw \in \mathtt{image}(\Gamma) + \epsilon, \end{align*}\]

where

\[ \sigma R = \{ R^{-1}B \mid B \in \mathcal{B}(\mathbb{R}^m) \} \]

and $E(-, w)$ is the negative log-density of the multivariate normal distribution $\mathcal{N}(Rw, \Gamma)$.

If $\mu \in \mathtt{image}(R : \Gamma)$, then System(R, ϵ) constructs the Gaussian system $\Sigma = (\mathbb{R}^n, \sigma R, P)$, where $P$ is the distribution of $\hat{w}$.

In particular, if $R$ has full row-rank, then $Rw = \epsilon$ is a kernel representation of $\Sigma$.

AlgebraicInference.System — Method

System(R::AbstractMatrix)

Let $R$ be an $m \times n$ matrix. Then System(R) constructs the deterministic Gaussian system $\Sigma = (\mathbb{R}^n, \sigma R, P),$ where

\[ \sigma R = \{ R^{-1}B \mid B \in \mathcal{B}(\mathbb{R}^m)\}\]

and

\[ P(R^{-1}B) = \begin{cases} 1 & 0 \in B \\ 0 & \text{else} \end{cases}.\]

Base.length — Method

length(Σ::AbstractSystem)

Let $\Sigma = (\mathbb{R}^n, \mathcal{E}, P)$. Then length(Σ) gets the dimension $n$.

StatsAPI.dof — Method

dof(Σ::AbstractSystem)

Get the number of degrees of freedom of $Σ$.

AlgebraicInference.fiber — Method

fiber(Σ::AbstractSystem)

Compute a basis for the fiber of $Σ$.

Statistics.mean — Method

mean(Σ::AbstractSystem)

Let $Rw = \epsilon$ be any kernel representation of $\Sigma$. Then mean(Σ) computes a vector $\mu$ such that $R\mu$ is the mean of $\epsilon$.

In particular, if $\Sigma$ is a classical Gaussian system, then $\mu$ is the mean of $\Sigma$.

Statistics.cov — Method

cov(Σ::AbstractSystem)

Let $Rw = \epsilon$ be any kernel representation of $\Sigma$. Then cov(Σ) computes a matrix $\Gamma$ such that $R \Gamma R^\mathsf{T}$ is the covariance of $\epsilon$.

In particular, if $\Sigma$ is a classical Gaussian system, then $\Gamma$ is the covariance of $\Sigma$.

Base.:* — Method

*(M::AbstractMatrix, Σ::AbstractSystem)

Let $M$ be an $n \times m$ matrix, and let $\Sigma = (\mathbb{R}^m, \mathcal{E}, P)$. Then M * Σ computes the Gaussian system $\Sigma' = (\mathbb{R}^n, \mathcal{E}', P')$, where

\[ \mathcal{E}' = \{ B \in \mathcal{B}(\mathbb{R}^n) \mid M^{-1}B \in \mathcal{E} \}\]

and

\[ P'(B) = P(M^{-1}B).\]

Base.:\ — Method

\(M::AbstractMatrix, Σ::AbstractSystem)

Catlab.Theories.:⊗ — Method

⊗(Σ₁::AbstractSystem, Σ₂::AbstractSystem)

Compute the product $\Sigma_1 \times \Sigma_2$.

Catlab.WiringDiagrams.WiringDiagramAlgebras.oapply — Method

oapply(wd::UndirectedWiringDiagram, box_map::AbstractDict{<:Any, <:AbstractSystem})

See oapply(wd::UndirectedWiringDiagram, box_map::AbstractVector{<:AbstractSystem}).

Catlab.WiringDiagrams.WiringDiagramAlgebras.oapply — Method

oapply(wd::UndirectedWiringDiagram, boxes::AbstractVector{<:AbstractSystem})

Compose Gaussian systems according to the undirected wiring diagram wd.

Valuations

AlgebraicInference.Valuation — Type

Valuation{T}

Abstract type for valuations in a valuation algebra.

Subtypes should specialize the following methods:

domain(ϕ::Valuation)

combine(ϕ₁::Valuation, ϕ₂::Valuation)

project(ϕ::Valuation, x)

Valuations are parametrized by the type of the variables in their variable system. If ϕ <: Valuation{T}, then domain(ϕ) should return a container with element type T.

References:

Pouly, M.; Kohlas, J. Generic Inference. A Unified Theory for Automated Reasoning; Wiley: Hoboken, NJ, USA, 2011.

AlgebraicInference.IdentityValuation — Type

IdentityValuation{T} <: Valuation{T}

The identity element $e$.

AlgebraicInference.LabeledBox — Type

LabeledBox{T₁, T₂} <: Valuation{T₁}

AlgebraicInference.domain — Function

domain(ϕ::Valuation)

Get the domain of $\phi$.

AlgebraicInference.combine — Function

combine(ϕ₁::Valuation, ϕ₂::Valuation)

Perform the combination $\phi_1 \otimes \phi_2$.

AlgebraicInference.project — Function

project(ϕ::Valuation, x)

Perform the projection $\phi^{\downarrow x}$.

AlgebraicInference.inference_problem — Method

inference_problem(wd::UndirectedWiringDiagram, box_map::AbstractDict)

See inference_problem(wd::UndirectedWiringDiagram, boxes::AbstractVector).

AlgebraicInference.inference_problem — Method

inference_problem(wd::UndirectedWiringDiagram, boxes::AbstractVector)

Let $f$ be an operation in Cospan of the form

\[ B \xleftarrow{\mathtt{box}} P \xrightarrow{\mathtt{junc}} J \xleftarrow{\mathtt{junc'}} Q\]

and $(b_1, \dots, b_n)$ a sequence of fillers for the boxes in $f$. Then inference_problem(wd, boxes) constructs a knowledge base $\{\phi_1, \dots, \phi_n\}$ and query $x \subseteq J$ such that

\[ (\phi_1 \otimes \dots \otimes \phi_n)^{\downarrow x} \cong F(f)(b_1, \dots, b_n),\]

where $F$ is the Cospan-algebra computed by oapply.

The operation $f$ must satify must satisfy the following constraints:

$\mathtt{junc'}$ is injective.

$\mathtt{image}(\mathtt{junc'}) \subseteq \mathtt{image}(\mathtt{junc})$

For all $x, y \in P$, $\mathtt{box}(x) = \mathtt{box}(y)$ and $\mathtt{junc}(x) = \mathtt{junc}(y)$ implies that $x = y$.