Library Reference

Systems

AlgebraicInference.GaussianSystem — Type

GaussianSystem{
    T₁ <: AbstractMatrix,
    T₂ <: AbstractMatrix, 
    T₃ <: AbstractVector,
    T₄ <: AbstractVector}

A Gaussian system.

AlgebraicInference.GaussianSystem — Method

GaussianSystem{T₁, T₂, T₃, T₄}(P, S, p, s, σ) where {
    T₁ <: AbstractMatrix,
    T₂ <: AbstractMatrix,
    T₃ <: AbstractVector,
    T₄ <: AbstractVector}

Construct a Gaussian system by specifying its energy function.

You should set σ equal to $s^\mathsf{T} S^+ s$, where $S^+$ is the Moore-Penrose psuedoinverse of $S$.

GaussianSystem(
    P::AbstractMatrix,
    S::AbstractMatrix,
    p::AbstractVector,
    s::AbstractVector,
    σ)

Construct a Gaussian system by specifying its energy function.

You should set σ equal to $s^\mathsf{T} S^+ s$, where $S^+$ is the Moore-Penrose psuedoinverse of $S$.

AlgebraicInference.normal — Function

normal(μ::AbstractVector, Σ::AbstractMatrix)

Construct a multivariate normal distribution with mean vector μ and covariance matrix Σ.

normal(μ::Real, σ::Real)

Construct a normal distribution with mean μ and standard deviation σ.

AlgebraicInference.kernel — Function

kernel(L::AbstractMatrix, μ::AbstractVector, Σ::AbstractMatrix)

Construct a conditional distribution of the form $(y \mid x) \sim \mathcal{N}(Lx + \mu, \Sigma).$

kernel(l::AbstractVector, μ::Real, σ::Real)

Construct a conditional distribution of the form $(y \mid x) \sim \mathcal{N}(l^\mathsf{T}x + \mu, \sigma^2).$

Base.length — Method

length(Σ::GaussianSystem)

Get the dimension of Σ.

Statistics.cov — Method

cov(Σ::GaussianSystem)

Get the covariance matrix of Σ.

AlgebraicInference.invcov — Method

invcov(Σ::GaussianSystem)

Get the precision matrix of Σ.

Statistics.var — Method

var(Σ::GaussianSystem)

Get the variances of Σ.

Statistics.mean — Method

mean(Σ::GaussianSystem)

Get the mean vector of Σ.

Catlab.WiringDiagrams.WiringDiagramAlgebras.oapply — Method

oapply(wd::AbstractUWD, systems::AbstractVector{<:GaussianSystem})

Compose Gaussian systems according to the undirected wiring diagram wd.

Problems

AlgebraicInference.InferenceProblem — Type

InferenceProblem{T₁, T₂ <: Union{Nothing, AbstractVector}}

An inference problem over a valuation algebra. Construct a solver for an inference problem with the function init, or solve it directly with solve.

AlgebraicInference.MinDegree — Type

MinDegree

Contructs a covering join tree for an inference problem using the variable elimination algorithm. Variables are eliminated according to the "minimum degree" heuristic.

AlgebraicInference.MinFill — Type

MinFill

Contructs a covering join tree for an inference problem using the variable elimination algorithm. Variables are eliminated according to the "minimum fill" heuristic.

AlgebraicInference.InferenceProblem — Method

InferenceProblem{T}(wd::AbstractUWD, hom_map::AbstractDict,
    ob_map::Union{Nothing, AbstractDict}=nothing;
    hom_attr=:name, ob_attr=:variable) where T

Construct an inference problem that performs undirected composition. Before being composed, the values of hom_map are converted to type T.

AlgebraicInference.InferenceProblem — Method

InferenceProblem{T}(wd::AbstractUWD, homs::AbstractVector,
    obs::Union{Nothing, AbstractVector}=nothing) where T

Construct an inference problem that performs undirected composition. Before being composed, the elements of homs are converted to type T.

CommonSolve.solve — Method

solve(ip::InferenceProblem, alg)

Solve an inference problem. The options for alg are

MinDegree()

MinFill()

CommonSolve.init — Method

init(ip::InferenceProblem, alg)

Construct a solver for an inference problem. The options for alg are

MinDegree()

MinFill()

Solvers

AlgebraicInference.InferenceSolver — Type

InferenceSolver{T₁, T₂ <: Union{Nothing, AbstractVector}}

This is the type constructed by init(ip::InferenceProblem). Use it with solve or solve! to solve inference problems.

An InferenceSolver can be reused to answer multiple queries:

is = init(ip)
sol1 = solve(is)
is.query = query2
sol2 = solve(is)

CommonSolve.solve — Method

solve(is::InferenceSolver)

Solve an inference problem.

CommonSolve.solve! — Method

solve!(is::InferenceSolver)

Solve an inference problem, caching intermediate computations.