Library Reference

Systems

AlgebraicInference.GaussianSystem — Type

GaussianSystem{
    T₁ <: AbstractMatrix,
    T₂ <: AbstractMatrix, 
    T₃ <: AbstractVector,
    T₄ <: AbstractVector}

A Gaussian system.

AlgebraicInference.GaussianSystem — Method

GaussianSystem{T₁, T₂, T₃, T₄}(P, S, p, s, σ) where {
    T₁ <: AbstractMatrix,
    T₂ <: AbstractMatrix,
    T₃ <: AbstractVector,
    T₄ <: AbstractVector}

Construct a Gaussian system by specifying its energy function.

You should set σ equal to $s^\mathsf{T} S^+ s$, where $S^+$ is the Moore-Penrose psuedoinverse of $S$.

GaussianSystem(
    P::AbstractMatrix,
    S::AbstractMatrix,
    p::AbstractVector,
    s::AbstractVector,
    σ)

Construct a Gaussian system by specifying its energy function.

You should set σ equal to $s^\mathsf{T} S^+ s$, where $S^+$ is the Moore-Penrose psuedoinverse of $S$.

AlgebraicInference.normal — Function

normal(Σ::AbstractMatrix, μ::AbstractVector)

Construct a multivariate normal distribution with covariance matrix Σ and mean vector μ.

AlgebraicInference.kernel — Function

kernel(Σ::AbstractMatrix, μ::AbstractVector, L::AbstractMatrix)

Construct a conditional distribution of the form

\[ y \mid x \sim \mathcal{N}(Lx + \mu, \Sigma).\]

Base.length — Method

length(Σ::GaussianSystem)

Get the dimension of Σ.

Statistics.cov — Method

cov(Σ::GaussianSystem)

Get the covariance matrix of Σ.

AlgebraicInference.invcov — Method

invcov(Σ::GaussianSystem)

Get the precision matrix of Σ.

Statistics.var — Method

var(Σ::GaussianSystem)

Get the variances of Σ.

Statistics.mean — Method

mean(Σ::GaussianSystem)

Get the mean vector of Σ.

Catlab.WiringDiagrams.WiringDiagramAlgebras.oapply — Method

oapply(wd::AbstractUWD, systems::AbstractVector{<:GaussianSystem})

Compose Gaussian systems according to the undirected wiring diagram wd.

Problems

AlgebraicInference.InferenceProblem — Type

InferenceProblem{T₁ <: Valuation, T₂}

An inference problem over a valuation algebra. Construct a solver for an inference problem with the function init, or solve it directly with solve.

AlgebraicInference.UWDProblem — Type

UWDProblem{T₁, T₂} = InferenceProblem{UWDBox{T₁, T₂}, T₂}

An inference problem that performs undirected composition.

AlgebraicInference.MinWidth — Type

MinWidth

Contructs a covering join tree for an inference problem using the variable elimination algorithm. Variables are eliminated according to the "min-width" heuristic.

AlgebraicInference.MinFill — Type

MinFill

Contructs a covering join tree for an inference problem using the variable elimination algorithm. Variables are eliminated according to the "min-fill" heuristic.

AlgebraicInference.UWDProblem — Method

UWDProblem{T}(wd::AbstractUWD, hom_map::AbstractDict,
    ob_map::Union{Nothing, AbstractDict}=nothing;
    hom_attr=:name, ob_attr=:variable) where T

Construct an inference problem that performs undirected composition. Before being composed, the values of hom_map are converted to type T.

UWDProblem{T}(wd::AbstractUWD, homs, obs=nothing) where T

Construct an inference problem that performs undirected composition. Before being composed, the elements of homs are converted to type T.

AlgebraicInference.UWDProblem — Method

UWDProblem{T}(wd::AbstractUWD, homs, obs=nothing) where T

Construct an inference problem that performs undirected composition. Before being composed, the elements of homs are converted to type T.

CommonSolve.solve — Method

solve(ip::InferenceProblem, alg)

Solve an inference problem. The options for alg are

MinWidth()

MinFill()

CommonSolve.init — Method

init(ip::InferenceProblem, alg)

Construct a solver for an inference problem. The options for alg are

MinWidth()

MinFill()

Solvers

AlgebraicInference.InferenceSolver — Type

InferenceSolver{T₁, T₂}

This is the type constructed by init(ip::InferenceProblem). Use it with solve or solve! to solve inference problems.

An InferenceSolver can be reused to answer multiple queries:

is = init(ip)
sol1 = solve(is)
is.query = query2
sol2 = solve(is)

AlgebraicInference.UWDSolver — Type

UWDSolver{T₁, T₂} = InferenceSolver{UWDBox{T₁, T₂}, T₂}

Solves a UWDProblem.

CommonSolve.solve — Method

solve(is::InferenceSolver)

Solve an inference problem.

CommonSolve.solve! — Method

solve!(is::InferenceSolver)

Solve an inference problem, caching intermediate computations.