Library Reference

Systems

AlgebraicInference.GaussianSystem — Type

GaussianSystem{T₁, T₂, T₃, T₄, T₅}

A Gaussian system.

AlgebraicInference.CanonicalForm — Type

CanonicalForm{T₁, T₂}(K, h) where {T₁ <: AbstractMatrix, T₂ <: AbstractVector}

Construct the canonical form $\mathcal{C}(K, h, g)$, where the normalization constant $g$ is inferred from $K$ and $h$.

CanonicalForm(K::AbstractMatrix, h::AbstractVector)

Construct the canonical form $\mathcal{C}(K, h, g)$, where the normalization constant $g$ is inferred from $K$ and $h$.

AlgebraicInference.GaussianSystem — Method

GaussianSystem(
    P::AbstractMatrix,
    S::AbstractMatrix,
    p::AbstractVector,
    s::AbstractVector,
    σ::Real)

Construct a Gaussian system by specifying its energy function.

You should set σ equal to $s^\mathsf{T} S^+ s$, where $S^+$ is the Moore-Penrose psuedoinverse of $S$.

AlgebraicInference.CanonicalForm — Method

CanonicalForm{T₁, T₂}(K, h) where {T₁ <: AbstractMatrix, T₂ <: AbstractVector}

Construct the canonical form $\mathcal{C}(K, h, g)$, where the normalization constant $g$ is inferred from $K$ and $h$.

CanonicalForm(K::AbstractMatrix, h::AbstractVector)

Construct the canonical form $\mathcal{C}(K, h, g)$, where the normalization constant $g$ is inferred from $K$ and $h$.

AlgebraicInference.normal — Function

normal(μ::AbstractVector, Σ::AbstractMatrix)

Construct a multivariate normal distribution with mean vector μ and covariance matrix Σ.

normal(μ::Real, σ::Real)

Construct a normal distribution with mean μ and standard deviation σ.

AlgebraicInference.kernel — Function

kernel(L::AbstractMatrix, μ::AbstractVector, Σ::AbstractMatrix)

Construct a conditional distribution of the form $(y \mid x) \sim \mathcal{N}(Lx + \mu, \Sigma).$

kernel(l::AbstractVector, μ::Real, σ::Real)

Construct a conditional distribution of the form $(y \mid x) \sim \mathcal{N}(l^\mathsf{T}x + \mu, \sigma^2).$

Base.length — Method

length(Σ::GaussianSystem)

Get the dimension of Σ.

Statistics.cov — Method

cov(Σ::GaussianSystem; atol::Real=1e-8)

Get the covariance matrix of Σ.

Distributions.invcov — Method

invcov(Σ::GaussianSystem)

Get the precision matrix of Σ.

Statistics.var — Method

var(Σ::GaussianSystem; atol::Real=1e-8)

Get the variances of Σ.

Statistics.mean — Method

mean(Σ::GaussianSystem; atol::Real=1e-8)

Get the mean vector of Σ.

Catlab.WiringDiagrams.WiringDiagramAlgebras.oapply — Method

oapply(wd::AbstractUWD, homs::AbstractVector{<:GaussianSystem}, obs::AbstractVector)

Compose Gaussian systems according to the undirected wiring diagram wd.

Problems

AlgebraicInference.InferenceProblem — Type

InferenceProblem{T₁, T₂, T₃, T₄}

An inference problem computes the conditional distribution of $X$ given $Y = e$, where $X$ and $Y$ are random variables whose joint probability distribution is specified by a graphical model.

AlgebraicInference.InferenceProblem — Method

InferenceProblem(
    uwd::AbstractUWD,
    hom_map::AbstractDict,
    ob_map::AbstractDict,
    evidence::AbstractDict=Dict();
    hom_attr::Symbol=:name,
    ob_attr::Symbol=:variable)

Construct an inference problem that performs undirected compositon.

AlgebraicInference.InferenceProblem — Method

InferenceProblem(
    uwd::AbstractUWD,
    homs::AbstractVector,
    obs::AbstractVector,
    evidence::AbstractDict=Dict())

Construct an inference problem that performs undirected compositon.

AlgebraicInference.InferenceProblem — Method

InferenceProblem(
    bn::BayesNet,
    query::AbstractVector,
    evidence::AbstractDict=Dict())

Construct an inference problem that queries a Bayesian network.

CommonSolve.solve — Method

solve(ip::InferenceProblem, alg::EliminationAlgorithm)

Solve an inference problem.

CommonSolve.init — Method

init(ip::InferenceProblem, alg::EliminationAlgorithm)

Construct a solver for an inference problem.

Solvers

AlgebraicInference.InferenceSolver — Type

InferenceSolver{T₁, T₂, T₃}

A solver for an inference problem.

An InferenceSolver can be reused to answer multiple queries:

is = init(ip)
sol₁ = solve(is)
is.query = query
sol₂ = solve(is)

CommonSolve.solve! — Method

solve!(is::InferenceSolver)

Solve an inference problem, caching intermediate computations.

Algorithms

AlgebraicInference.EliminationAlgorithm — Type

EliminationAlgorithm

An algorithm for computing an elimination order for an undirected graph.

AlgebraicInference.MinDegree — Type

MinDegree <: EliminationAlgorithm

The minimum-degree heuristic.

AlgebraicInference.MinFill — Type

MinFill <: EliminationAlgorithm

The minimum-fill heuristic.

AlgebraicInference.CuthillMcKeeJL_RCM — Type

CuthillMcKeeJL_RCM <: EliminationAlgorithm

The reverse Cuthill-McKee algorithm. Uses CuthillMckee.jl.

AlgebraicInference.AMDJL_AMD — Type

AMDJL_AMD <: EliminationAlgorithm

The approximate minimum degree algorithm. Uses AMD.jl.

AlgebraicInference.MetisJL_ND — Type

MetisJL_ND <: EliminationAlgorithm

The nested dissection heuristic. Uses Metis.jl.